Crate kernel

Modules

arch
RISC-V related functions

clint
RISC-V Core Local Interrupter

elf
ELF parsing

file
File in core-os including file in filesystem, device, pipe and symbol link

intr
Handle interrupts

jump
Never type and jump after RAII

mem
Allocator implementation

page
Paging implementaion and page table abstraction

plic
RISC-V Platform-Level Interrupt Controller

print
Macros for output

process
Process and scheduling

sleeplock
A lock which puts current process into sleep and allows interrupt

spinlock
A RISC-V Mutex

start
RISC-V hart boot and initialize

symbols
Linker-script constants and RISC-V constants (This module may be viewed as memory layout definitions)

syscall
Module for processing syscall

test
Test runner

trap
Machine mode and supervisor mode traps

uart
UART driver module

virtio
virt-io driver

Macros

info
Prints an info, with newline.

panic_println
Create a new UART object and print

print
Print information

println
Print with a new line

warn
Prints a warning, with newline.

Functions

abort
Abort function

eh_personality

panic
Panic handler