kernel::acpi::aml::parser

Struct Parser

pub struct Parser<'a> {
    code: &'a [u8],
    pos: usize,
    state: State<'a>,
}

Fields
code: &'a [u8]pos: usizestate: State<'a>
Implementations

impl Parser<'_>

fn remaining_bytes(&self) -> usize

fn get_next_byte(&mut self) -> Result<u8, AmlParseError>

fn peek_next_byte(&self) -> Result<u8, AmlParseError>

fn forward(&mut self, n: usize) -> Result<(), AmlParseError>

fn backward(&mut self, n: usize) -> Result<(), AmlParseError>

fn get_pkg_length(&mut self) -> Result<usize, AmlParseError>

fn get_inner_parser(&mut self) -> Result<Parser<'_>, AmlParseError>

fn clone_parser(&mut self) -> Parser<'_>

Renamed to not be confused with Clone::clone

fn check_empty(&self) -> Result<(), AmlParseError>

fn parse_term(&mut self) -> Result<AmlTerm, AmlParseError>

fn predict_possible_args( &mut self, expect_data_after: bool, name: &str, ) -> usize

fn try_parse_term( &mut self, opcode: u8, ) -> Result<Option<AmlTerm>, AmlParseError>

fn parse_term_arg_non_method_arg(&mut self) -> Result<TermArg, AmlParseError>

similar to Self::parse_term_arg, but cannot call methods, as in some places method calls are not allowed

TODO: This should be removed, as in general a method call is a valid term arg, its just we break some parts due to us not knowing if a name is a method or not, and prediction predicts wrong and messes up This happens for + and >> and <<, cases I have seen and know of bugs in the parsing

fn parse_term_arg_last(&mut self) -> Result<TermArg, AmlParseError>

similar to Self::parse_term_arg, but doesn’t expect to have data after it, i.e. last in statements or something similar

fn parse_term_arg(&mut self) -> Result<TermArg, AmlParseError>

fn parse_package_element( &mut self, ) -> Result<PackageElement<UnresolvedDataObject>, AmlParseError>

fn parse_eisa_id(id: u32) -> String

fn try_parse_data_object( &mut self, ) -> Result<Option<UnresolvedDataObject>, AmlParseError>

fn parse_term_arg_general( &mut self, can_call_method: bool, expect_data_after: bool, ) -> Result<TermArg, AmlParseError>

fn try_parse_name(&mut self) -> Result<Option<String>, AmlParseError>

fn parse_name(&mut self) -> Result<String, AmlParseError>

fn try_parse_local(&mut self, lead: u8) -> Result<Option<u8>, AmlParseError>

fn try_parse_arg(&mut self, lead: u8) -> Result<Option<u8>, AmlParseError>

fn parse_target(&mut self) -> Result<Box<Target>, AmlParseError>

fn parse_term_list(&mut self) -> Result<Vec<AmlTerm>, AmlParseError>

fn parse_fields_list_and_flags( self, ) -> Result<(u8, Vec<FieldElement>), AmlParseError>

fn parse_root(&mut self) -> Result<AmlCode, AmlParseError>