kernel::acpi

Module tables

Re-exports

pub use facp::Facp;

Modules

facp

Structs

Apic

ApicGenericAddress

Bgrt

BiosTables

DescriptionHeader

DescriptorTable

GenericInitiatorAffinity

GicInterruptTranslationServiceAffinity

GiccAffinity

Hpet

InterruptSourceOverride

IoApic

LocalApicAddressOverride

LocalApicNmi

MemoryAffinity

NonMaskableInterrupt

ProcessorLocalAcpiAffinity

ProcessorLocalApic

ProcessorLocalX2ApicAffinity

Rsdp

Represent v2 and above

RsdpV1

RsdpV2

Rsdt

Srat

Waet

Xsdt

This is inside DSDT and SSDT

Enums

DescriptorTableBody

InterruptControllerStruct

StaticResourceAffinity

Constants

BIOS_RO_MEM_END 🔒

BIOS_RO_MEM_START 🔒

Statics

BIOS_TABLES 🔒

Functions

get_acpi_table_bytes 🔒 ⚠

Safety

get_acpi_tables

get_struct_from_bytes 🔒

Will fill the table from the header data, and zero out remaining bytes if any are left

get_table_from_body 🔒 ⚠

Will fill the table from the header data, and zero out remaining bytes if any are left

init_acpi_tables