Struct kernel::arch::page::PageErrorCode

pub struct PageErrorCode(/* private fields */);

Implementations

impl PageErrorCode

pub const PRESENT: Self = _

When set, the page fault was caused by a page-protection violation. When not set, it was caused by a non-present page.

pub const WRITE: Self = _

When set, the page fault was caused by a write access. When not set, it was caused by a read access.

pub const USER: Self = _

When set, the page fault was caused while CPL = 3. This does not necessarily mean that the page fault was a privilege violation.

pub const RESERVED_BITS: Self = _

When set, one or more page directory entries contain reserved bits which are set to 1. This only applies when the PSE or PAE flags in CR4 are set to 1.

pub const FETCH: Self = _

When set, the page fault was caused by an instruction fetch. This only applies when the No-Execute bit is supported and enabled.

pub const PROTECTION_KEY: Self = _

When set, the page fault was caused by a protection-key violation. The PKRU register (for user-mode accesses) or PKRS MSR (for supervisor-mode accesses) specifies the protection key rights.

pub const SHADOW_STACK: Self = _

When set, the page fault was caused by a shadow stack access.

pub const SGX: Self = _

When set, the fault was due to an SGX violation. The fault is unrelated to ordinary paging.

impl PageErrorCode

pub const fn from_bits_retain(bits: u16) -> Self

pub const fn from_bits_truncate(bits: u16) -> Self

pub const fn bits(&self) -> u16

pub const fn empty() -> Self

pub const fn is_empty(&self) -> bool

pub const fn all() -> Self

pub const fn is_all(&self) -> bool

pub const fn contains(&self, other: Self) -> bool

pub const fn insert(&mut self, other: Self)

pub const fn remove(&mut self, other: Self)

pub const fn toggle(&mut self, other: Self)

pub const fn set(&mut self, other: Self, value: bool)

pub const fn intersects(&self, other: Self) -> bool

pub const fn intersection(self, other: Self) -> Self

pub const fn union(self, other: Self) -> Self

pub const fn difference(self, other: Self) -> Self

pub const fn symmetric_difference(self, other: Self) -> Self

pub const fn complement(self) -> Self

impl PageErrorCode

pub fn is_page_present(&self) -> bool

pub fn could_not_read(&self) -> bool

pub fn could_not_write(&self) -> bool

pub fn could_not_execute(&self) -> bool