Struct kernel::bus::pci::PciClass

pub struct PciClass(_);

A type that defines the PCI class code and interface.

For example, the class code for XHCI (0x0C_03_30) is expressed as follows.

let cc = PciClass::code(0x0C).sub(0x03).interface(0x30);

To see if one class code is included in another class code or interface, compare the following

let mask = PciClass::code(0x03).sub(0x00);
if cc.matches(mask) {
  // code here
}

Implementations

impl PciClass

pub const fn from_pci(data: u32) -> Self

Makes an instance from the PCI class code register (0x02)

pub const fn code(code: u8) -> Self

Makes an instance from class code.

pub const fn sub(self, sub: u8) -> Self

Chains subclasses to the class code.

pub const fn interface(self, interface: u8) -> Self

Chains the programming interface to the class code and subclasses.

pub const fn raw_data(&self) -> u32

pub const fn matches(&self, other: Self) -> bool

Returns whether or not this instance matches the specified class code, subclass, or programming interface.