Module kernel::c_types

C types for the bindings.

The bindings generated by bindgen use these types to map to the C ones.

C’s standard integer types may differ in width depending on the architecture, thus we need to conditionally compile those.

Type Definitions

c_char

C char type.

c_int

C int type.

c_long

C long type.

c_longlong

C long long type.

c_schar

C signed char type.

c_short

C short type.

c_size_t

C size_t type (typically defined in <stddef.h>).

c_ssize_t

C ssize_t type (typically defined in <sys/types.h> by POSIX).

c_uchar

C unsigned char type.

c_uint

C unsigned int type.

c_ulong

C unsigned long type.

c_ulonglong

C unsigned long long type.

c_ushort

C unsigned short type.

c_void

C void type.