Struct kernel::chrdev::Registration

pub struct Registration<const N: usize> { /* fields omitted */ }

Character device registration.

May contain up to a fixed number (N) of devices. Must be pinned.

Implementations

impl<const N: usize> Registration<{ N }>

pub fn new(
name: &'static CStr,
minors_start: u16,
this_module: &'static ThisModule
) -> Self

Creates a Registration object for a character device.

This does not register the device: see Self::register().

This associated function is intended to be used when you need to avoid a memory allocation, e.g. when the Registration is a member of a bigger structure inside your crate::KernelModule instance. If you are going to pin the registration right away, call Self::new_pinned() instead.

pub fn new_pinned(
name: &'static CStr,
minors_start: u16,
this_module: &'static ThisModule
) -> Result<Pin<Box<Self>>>

Creates a pinned Registration object for a character device.

This does not register the device: see Self::register().

pub fn register<T: FileOpener<()>>(self: Pin<&mut Self>) -> Result

Registers a character device.

You may call this once per device type, up to N times.