kernel

Module cpu

Re-exports

pub use rip;

pub use rbp;

pub use rsp;

Modules

cpuid

flags

gdt

idt

interrupts

msr

Structs

Cpu

Constants

MAX_CPUS ğŸ”’

Statics

CPUS ğŸ”’

Traits

IoPortInt

Functions

clear_interruptsâš

cpu

get_cr0âš

get_cr3âš

get_cr4âš

get_cs ğŸ”’

haltâš

inb ğŸ”’ âš

ind ğŸ”’ âš

invalidate_tlpâš

inw ğŸ”’ âš

io_inâš

io_outâš

lgdt ğŸ”’ âš

SAFETY: the data pointed to by gdtr must be static and never change

lidt ğŸ”’ âš

SAFETY: the data pointed to by ldtr must be static and never change

ltr ğŸ”’ âš

outb ğŸ”’ âš

outd ğŸ”’ âš

outw ğŸ”’ âš

read_tscâš

rflagsâš

set_cr0âš

set_cr3âš

set_cr4âš

set_cs ğŸ”’ âš

set_data_segments ğŸ”’ âš

set_interruptsâš