kernel::cpu

Module gdt

Modules

flags 🔒

Structs

GlobalDescriptorManager 🔒

GlobalDescriptorTable 🔒

GlobalDescriptorTablePointer 🔒

SegmentSelector

SystemDescriptorEntry 🔒

System Descriptor Entry for GDT

TaskStateSegment 🔒

Task State Segment

UserDescriptorEntry 🔒

User Descriptor Entry for GDT

Constants

KERNEL_RING

USER_RING

Statics

GDT 🔒

TSS 🔒

SAFETY: TSS is only used when GDT is locked, so its safe to use as static mut

Functions

get_user_code_seg_index

get_user_data_seg_index

init_kernel_gdt

This should be called only once, otherwise, it will crash