Module kernel::debug

Provides a debug! macro for in-kernel debugging.

This module uses an internal buffer to write the strings into. If you are writing and the buffer fills up, you can make the size of output_buffer larger.

Example

debug!("Yes the code gets here with value {}", i);

Reexports

use callback::AppId;

use callback::Callback;

use core::cmp::min;

use core::fmt::Arguments;

use core::fmt::Result;

use core::fmt::Write;

use core::fmt::write;

use core::ptr::read_volatile;

use core::ptr::write_volatile;

use core::str;

use driver::Driver;

use mem::AppSlice;

use returncode::ReturnCode;

Structs

DebugWriter

Constants

APPID_IDX

BUF_SIZE

Statics

DEBUG_WRITER

KERNEL_CONSOLE_CALLBACK

Traits

Debug

Functions

assign_console_driver⚠

begin_debug

begin_debug_fmt

get_container⚠