Struct kernel::device::Device

pub struct Device { /* fields omitted */ }

A ref-counted device.

Invariants

ptr is valid, non-null, and has a non-zero reference count. One of the references is owned by self, and will be decremented when self is dropped.

Implementations

impl Device

pub unsafe fn new(ptr: *mut device) -> Self

Creates a new device instance.

Safety

Callers must ensure that ptr is valid, non-null, and has a non-zero reference count.

pub fn from_dev(dev: &dyn RawDevice) -> Self

Creates a new device instance from an existing RawDevice instance.