Module kernel::devices::keyboard

Reading keyboard events from KeyboardReader

Modules

modifier

status 🔒

Structs

Key

Keyboard

KeyboardDevice

KeyboardDeviceCreator

This is just a helper definition, and should not be used directly in any file

Enums

KeyType

Constants

KEYBOARD_BUFFER_SIZE 🔒

Number of key events that can be buffered before being overwritten We are expecting interested readers to be fast, so we don’t need a very large buffer

KEYBOARD_DATA_PORT 🔒

KEYBOARD_INT_NUM 🔒

KEYBOARD_STATUS_PORT 🔒

KEY_PRESSED 🔒

Statics

KEYBOARD 🔒

Functions

get_keyboard_reader

get_modifier 🔒

get_toggle 🔒

init_keyboard

key_type_from_device 🔒

keyboard_interrupt_handler 🔒

Type Aliases

KeyboardReader