kernel::devices

Module pipe

Structs

InnerPipe 🔒

Pipe is a device that allows two processes to communicate with each other.

PipeSide

Represent one side of a pipe. Check create_pipe_pair for more details.

Functions

create_pipe_pair

Create a connected pipe pair. The first returned file is the read side of the pipe. The second returned file is the write side of the pipe.