Module kernel::driver

Generic support for drivers of different buses (e.g., PCI, Platform, Amba, etc.).

Each bus/subsystem is expected to implement DriverOps, which allows drivers to register using the Registration class.

Structs

IdArray

A zero-terminated device id array, followed by context data.

IdTable

A device id table.

Module

A kernel module that only registers the given driver on init.

Registration

The registration of a driver.

Traits

DeviceRemoval

Custom code within device removal.

DriverOps

A subsystem (e.g., PCI, Platform, Amba, etc.) that allows drivers to be written for it.

RawDeviceId

Conversion from a device id to a raw device id.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.