Trait kernel::driver::DeviceRemoval

pub trait DeviceRemoval {
    fn device_remove(&self);
}

Custom code within device removal.

Required Methods

fn device_remove(&self)

Cleans resources up when the device is removed.

This is called when a device is removed and offers implementers the chance to run some code that cleans state up.