Struct kernel::drivers::usb::UsbDevice

pub struct UsbDevice { /* private fields */ }

USB device instance type

Implementations

impl UsbDevice

pub fn addr(&self) -> UsbAddress

Gets the USB address of this device.

pub fn parent(&self) -> Option<UsbAddress>

Gets the USB address of the parent device.

pub fn route_string(&self) -> UsbRouteString

pub fn children(&self) -> impl Iterator<Item = UsbAddress>

pub fn uuid(&self) -> &Uuid

Gets the uuid of this device, if available.

pub fn protocol_speed(&self) -> usize

pub fn usb_version(&self) -> UsbVersion

pub fn vid(&self) -> UsbVendorId

Gets the vendor ID for this device.

pub fn pid(&self) -> UsbProductId

Gets the product ID for this device.

pub fn class(&self) -> UsbClass

Gets the device class of this device.

pub fn preferred_lang_id(&self) -> UsbLangId

pub fn is_configured(&self) -> bool

pub fn manufacturer_string(&self) -> Option<&str>

Gets the manufacturer’s string for this device if possible.

pub fn preferred_device_name(&self) -> Option<&str>

Gets the preferred name of the device. First, use the product name if available. Next, use the class name if available.

pub fn product_string(&self) -> Option<&str>

Gets the product name string for this device, if possible.

pub fn serial_number(&self) -> Option<&str>

Gets the serial number string of this device, if possible.

pub fn descriptor(&self) -> &UsbDeviceDescriptor

Gets the device descriptor for this device.

pub fn config_raw(&self) -> &[u8]

Gets raw configuration descriptors.

pub fn bos(&self) -> &UsbBinaryObjectStore

pub fn current_configuration(&self) -> &UsbConfiguration

pub fn configurations(&self) -> &[UsbConfiguration]