Struct kernel::drivers::usb::xhci::Xhci

pub struct Xhci { /* private fields */ }

Extensible Host Controller Interface

Many methods are made public for documentation purposes, but are not intended to be called from the outside.

Implementations

impl Xhci

pub fn registrar() -> Box<dyn PciDriverRegistrar>

pub fn get_device_context(&self, slot_id: SlotId) -> PhysicalAddress

pub fn set_device_context(&self, slot_id: SlotId, value: PhysicalAddress)

pub fn ring_a_doorbell(&self, slot_id: Option<SlotId>, dci: Option<DCI>)

pub fn ring_a_doorbell_async( &self, parent_slot_id: Option<SlotId>, slot_id: Option<SlotId>, dci: Option<DCI> ) -> Result<(), ()>

pub fn port_by_slot(&self, slot_id: SlotId) -> Option<PortId>

pub fn slot_by_port(&self, port_id: PortId) -> Option<SlotId>

pub fn port_by(&self, port_id: PortId) -> &PortRegisters

pub fn port_is_usb2(&self, port_id: PortId) -> bool

pub fn port_is_usb3(&self, port_id: PortId) -> bool

pub fn ports<'a>(&self) -> impl Iterator<Item = (PortId, &'a PortRegisters)>

pub fn usb2_ports<'a>( &self ) -> impl Iterator<Item = (PortId, &'a PortRegisters)>

pub fn usb3_ports<'a>( &self ) -> impl Iterator<Item = (PortId, &'a PortRegisters)>

pub fn input_context<'a>(&self, slot_id: SlotId) -> &'a mut InputContext

pub fn wait_cnr(&self, _: usize)

wait for CNR (Controller Not Ready)

pub fn ep_ring_index( &self, slot_id: Option<SlotId>, dci: Option<DCI> ) -> Option<usize>

pub fn alloc_ep_ring( &self, slot_id: Option<SlotId>, dci: Option<DCI> ) -> Option<NonNullPhysicalAddress>

pub fn allocate_crb<'a>(&'a self) -> Option<&'a mut CommandRequestBlock>

pub fn find_crb<'a>( &'a self, scheduled_trb: ScheduledTrb, state: Option<RequestState> ) -> Option<&'a mut CommandRequestBlock>

pub fn issue_trb( &self, crb: Option<&mut CommandRequestBlock>, trb: &Trb, slot_id: Option<SlotId>, dci: Option<DCI> )

pub fn execute_command( &self, trb: &Trb ) -> Result<TrbCommandCompletionEvent, TrbCommandCompletionEvent>

Issue trb command

pub async unsafe fn control_async( self: Arc<Self>, device: &HciDeviceContext, setup: UsbControlSetupData, transfer_mode: TransferDirection<u8> ) -> Result<usize, UsbError>

pub async unsafe fn transfer_async( self: Arc<Self>, device: &HciDeviceContext, dci: DCI, transfer_mode: TransferDirection<u8>, len: usize ) -> Result<usize, UsbError>

pub fn reset_endpoint( &self, slot_id: SlotId, dci: DCI ) -> Result<TrbCommandCompletionEvent, TrbCommandCompletionEvent>

pub fn configure_endpoint( &self, slot_id: SlotId, dci: DCI, ep_type: EpType, max_packet_size: usize, interval: u8, copy_dc: bool )

pub fn configure_hub2( &self, slot_id: SlotId, hub_desc: &Usb2HubDescriptor, is_mtt: bool ) -> Result<(), UsbError>

pub fn configure_hub3( &self, slot_id: SlotId, hub_desc: &Usb3HubDescriptor ) -> Result<(), UsbError>

pub async fn attach_child_device( self: Arc<Self>, hub: Arc<HciContext>, port: UsbHubPortNumber, speed: PSIV ) -> Result<UsbAddress, UsbError>

pub async fn attach_root_device( self: &Arc<Self>, port_id: PortId ) -> Option<UsbAddress>

pub fn set_max_packet_size( &self, slot_id: SlotId, max_packet_size: usize ) -> Result<(), ()>

pub fn process_event(&self)

pub async fn _attach_root_port(self: &Arc<Self>, port_id: PortId)

pub async fn _process_port_change( self: &Arc<Self>, port_id: PortId, force: bool )

pub fn focus_hub(self: &Arc<Self>, hub: Option<SlotId>)

pub fn unfocus_hub(self: &Arc<Self>, hub: Option<SlotId>)