Module kernel::error

Kernel errors.

C header: include/uapi/asm-generic/errno-base.h

Modules

code

Contains the C-compatible error codes.

Structs

Error

Generic integer kernel error.

Functions

to_result

Calls a kernel function that returns an integer error code on failure and converts the result to a Result.

Type Definitions

Result

A Result with an Error error type.