Struct kernel::error::Error

pub struct Error(/* private fields */);

Generic integer kernel error.

The kernel defines a set of integer generic error codes based on C and POSIX ones. These codes may have a more specific meaning in some contexts.

Invariants

The value is a valid errno (i.e. >= -MAX_ERRNO && < 0).

Implementations

impl Error

pub fn to_errno(self) -> c_int

Returns the kernel error code.

pub fn name(&self) -> Option<&'static CStr>

Returns a string representing the error, if one exists.