Module kernel::executor

The executor module provides an executor that schedules and runs both the kernel threads and the user threads.

Re-exports

pub use context::TrapContext;

pub use future::spawn_thread;

pub use future::yield_now;

pub use future::ControlFlow;

Modules

context 🔒

The context module provides a TrapContext struct that save and restore the context of a thread when an exception or interrupt occurs.

future 🔒

scheduler 🔒

Structs

SCHEDULER 🔒

Statics

LAZY 🔒

Functions

__stability 🔒

__static_ref_initialize 🔒

init

Initializes the stvec to the address of the _enter_kernel_space function, which is located at the beginning of the TRAMPOLINE page.

run_until_complete

Runs an event loop that executes all the tasks in the TASK_QUEUE until there are no more task left.

spawn 🔒