Module kernel::executor::future

Structs

YieldFuture 🔒

Enums

ControlFlow

The ControlFlow enum specifies the operation that the executor should execute on a thread prior to returning to user space.

Functions

_enter_kernel_space 🔒 ⚠

_enter_user_space 🔒 ⚠

spawn_thread

thread_loop 🔒

The thread_loop future represents the lifetime of a user thread.

yield_now

Opaque Types

🔒

🔒