Module kernel::file::flags

Flags associated with a File.

Constants

O_ACCMODE

Bitmask for access mode flags.

O_APPEND

File is opened in append mode.

O_ASYNC

Signal-driven I/O is enabled.

O_CLOEXEC

Close-on-exec flag is set.

O_CREAT

File was created if it didn’t already exist.

O_DIRECT

Direct I/O is enabled for this file.

O_DIRECTORY

File must be a directory.

O_DSYNC

Like O_SYNC except metadata is not synced.

O_EXCL

Ensure that this file is created with the open(2) call.

O_LARGEFILE

Large file size enabled (off64_t over off_t).

O_NDELAY

Also known as O_NDELAY.

O_NOATIME

Do not update the file last access time.

O_NOCTTY

File should not be used as process’s controlling terminal.

O_NOFOLLOW

If basename of path is a symbolic link, fail open.

O_NONBLOCK

File is using nonblocking I/O.

O_PATH

Used to obtain a path file descriptor.

O_RDONLY

File is read only.

O_RDWR

File can be both read and written.

O_SYNC

Write operations on this file will flush data and metadata.

O_TMPFILE

This file is an unnamed temporary regular file.

O_TRUNC

File should be truncated to length 0.

O_WRONLY

File is write only.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.