Trait kernel::file_operations::PointerWrapper

pub trait PointerWrapper<T> {
    fn into_pointer(self) -> *const T;

    unsafe fn from_pointer(ptr: *const T) -> Self;
}

Used to convert an object into a raw pointer that represents it.

It can eventually be converted back into the object. This is used to store objects as pointers in kernel data structures, for example, an implementation of FileOperations in struct file::private_data.

Required methods

fn into_pointer(self) -> *const T

Returns the raw pointer.

unsafe fn from_pointer(ptr: *const T) -> Self

Returns the instance back from the raw pointer.

Safety

The passed pointer must come from a previous call to PointerWrapper::into_pointer().

Loading content...