Struct kernel::fs::ext2::block::BaseSuperblock

pub struct BaseSuperblock {

    inode_count: u32,
    blocks_count: u32,
    rblocks_num: u32,
    pub blocks_unalloc: u32,
    pub inode_unalloc: u32,
    pub superblock_block: u32,
    block_size: u32,
    frag_size: u32,
    bgroup_bno: u32,
    bgroup_fno: u32,
    bgroup_ino: u32,
    last_mt: u32,
    last_wt: u32,
    mount_no: u16,
    mount_no_max: u16,
    ext2_sig: u16,
    fs_state: u16,
    err_handle: u16,
    minor: u16,
    last_fsck: u32,
    fsck_interval: u32,
    osid: u32,
    major: u32,
    pub uid: u16,
    pub gid: u16,
    extension: Option<ExtendedSuperblock>,

}

Superblock always take 1024 bytes with/without Extended block

Fields
inode_count: u32

Total number of inodes in file system

blocks_count: u32

Total number of blocks in file system

rblocks_num: u32

Number of blocks reserved for superuser (see offset 80)

blocks_unalloc: u32

Total number of unallocated blocks

inode_unalloc: u32

Total number of unallocated inodes

superblock_block: u32

Block number of the block containing the superblock (also the starting block number, NOT always zero.)

block_size: u32

log2 (block size) - 10. (In other words, the number to shift 1,024 to the left by to obtain the block size)

frag_size: u32

log2 (fragment size) - 10. (In other words, the number to shift 1,024 to the left by to obtain the fragment size)

bgroup_bno: u32

Number of blocks in each block group

bgroup_fno: u32

Number of fragments in each block group

bgroup_ino: u32

Number of inodes in each block group

last_mt: u32

Last mount time (in POSIX time)

last_wt: u32

Last written time (in POSIX time)

mount_no: u16

Number of times the volume has been mounted since its last consistency check (fsck)

mount_no_max: u16

Number of mounts allowed before a consistency check (fsck) must be done

ext2_sig: u16

Ext2 signature (0xef53), used to help confirm the presence of Ext2 on a volume

fs_state: u16

File system state (see below)

err_handle: u16

What to do when an error is detected (see below)

minor: u16

Minor portion of version (combine with Major portion below to construct full version field)

last_fsck: u32

POSIX time of last consistency check (fsck)

fsck_interval: u32

Interval (in POSIX time) between forced consistency checks (fsck)

osid: u32

Operating system ID from which the filesystem on this volume was created (see below)

major: u32

Major portion of version (combine with Minor portion above to construct full version field)

uid: u16

User ID that can use reserved blocks

gid: u16

Group ID that can use reserved blocks

extension: Option<ExtendedSuperblock>
Implementations

impl BaseSuperblock

pub fn sig(&self) -> u16

pub fn version(&self) -> (u32, u16)

pub fn bsize(&self) -> usize

pub fn inode_size(&self) -> u16

pub fn block_per_grp(&self) -> u32

pub fn inode_per_grp(&self) -> u32

pub fn inode_count(&self) -> u32

pub fn block_count(&self) -> u32

pub fn block_grp_count(&self) -> u32

pub fn set_extension(&mut self, extension: ExtendedSuperblock)

impl BaseSuperblock

pub fn into_boxed_slice(&self) -> Box<[u8]>