Struct kernel::fs::ext2::block::ExtendedSuperblock

pub struct ExtendedSuperblock {

    first_inode: u32,
    inode_size: u16,
    bgroup_superblock: u16,
    opt_features: u32,
    req_features: u32,
    ro_features: u32,
    fsid: [u8; 16],
    vol_name: [u8; 16],
    last_path: [u8; 64],
    compr: u32,
    prealloc_blocks_files: u8,
    prealloc_block_dir: u8,
    unused: u16,
    journ_id: [u8; 16],
    journ_inode: u32,
    journ_dev: u32,
    orphan_inode_lst: u32,

}

Present if Major >= 1 Bytes from 236 to 1023 aren’t used

Fields
first_inode: u32

First non-reserved inode in file system. (In versions < 1.0, this is fixed as 11)

inode_size: u16

Size of each inode structure in bytes. (In versions < 1.0, this is fixed as 128)

bgroup_superblock: u16

Block group that this superblock is part of (if backup copy)

opt_features: u32

Optional features present (features that are not required to read or write, but usually result in a performance increase. see below)

req_features: u32

Required features present (features that are required to be supported to read or write. see below)

ro_features: u32

Features that if not supported, the volume must be mounted read-only see below)

fsid: [u8; 16]

File system ID (what is output by blkid)

vol_name: [u8; 16]

Volume name (C-style string: characters terminated by a 0 byte)

last_path: [u8; 64]

Path volume was last mounted to (C-style string: characters terminated by a 0 byte)

compr: u32

Compression algorithms used (see Required features above)

prealloc_blocks_files: u8

Number of blocks to preallocate for files

prealloc_block_dir: u8

Number of blocks to preallocate for directories

unused: u16

(Unused)

journ_id: [u8; 16]

Journal ID (same style as the File system ID above)

journ_inode: u32

Journal inode

journ_dev: u32

Journal device

orphan_inode_lst: u32

Head of orphan inode list