Module kernel::fs::ext2::inode

Structs

Dentry

Inode

Like blocks, each inode has a numerical address. It is extremely important to note that unlike block addresses, inode addresses start at 1. With Ext2 versions prior to Major version 1, inodes 1 to 10 are reserved and should be in an allocated state. Starting with version 1, the first non-reserved inode is indicated via a field in the Superblock. Of the reserved inodes, number 2 subjectively has the most significance as it is used for the root directory. Inodes have a fixed size of either 128 for version 0 Ext2 file systems, or as dictated by the field in the Superblock for version 1 file systems. All inodes reside in inode tables that belong to block groups. Therefore, looking up an inode is simply a matter of determining which block group it belongs to and indexing that block group’s inode table.

Enums

Dtype

Constants

IFLAG_ASDIR

IFLAG_FILECOMPR

IFLAG_HASHINDEX

IFLAG_IMMUTABLE

IFLAG_JOURN

IFLAG_KEEPCPY

IFLAG_NODUMP

IFLAG_NOUPDATE

IFLAG_OAPPEN

IFLAG_SECDEL

IFLAG_SYNC

IPERM_GEXEC

IPERM_GREAD

IPERM_GWRIT

IPERM_OEXEC

IPERM_OREAD

IPERM_OWRIT

IPERM_SETGID

IPERM_SETUID

IPERM_STICK

IPERM_UEXEC

IPERM_UREAD

IPERM_UWRIT

ITYPE_BLOCK

ITYPE_CHARDEV

ITYPE_DIR

ITYPE_FIFO

ITYPE_REGU

ITYPE_SOCK

ITYPE_SYMF