Struct kernel::fs::ext2::inode::Inode

pub struct Inode {

    pub tperm: u16,
    pub uid: u16,
    pub size_lh: u32,
    pub lat: u32,
    pub creatt: u32,
    pub lmt: u32,
    pub delt: u32,
    pub gid: u16,
    pub count_hl: u16,
    pub count_ds: u32,
    pub flags: u32,
    os_specific1: u32,
    pub dbp: [u32; 12],
    pub sibp: u32,
    pub dibp: u32,
    pub tibp: u32,
    pub gen_no: u32,
    pub facl: u32,
    pub size_uh: u32,
    pub block_addr: u32,
    os_specific2: [u8; 12],

}

Like blocks, each inode has a numerical address. It is extremely important to note that unlike block addresses, inode addresses start at 1. With Ext2 versions prior to Major version 1, inodes 1 to 10 are reserved and should be in an allocated state. Starting with version 1, the first non-reserved inode is indicated via a field in the Superblock. Of the reserved inodes, number 2 subjectively has the most significance as it is used for the root directory. Inodes have a fixed size of either 128 for version 0 Ext2 file systems, or as dictated by the field in the Superblock for version 1 file systems. All inodes reside in inode tables that belong to block groups. Therefore, looking up an inode is simply a matter of determining which block group it belongs to and indexing that block group’s inode table.

Fields
tperm: u16

Type and Permissions (see below)

uid: u16

User ID

size_lh: u32

Lower 32 bits of size in bytes

lat: u32

Last Access Time (in POSIX time)

creatt: u32

Creation Time (in POSIX time)

lmt: u32

Last Modification time (in POSIX time)

delt: u32

Deletion time (in POSIX time)

gid: u16

Group ID

count_hl: u16

Count of hard links (directory entries) to this inode. When this reaches 0, the data blocks are marked as unallocated.

count_ds: u32

Count of disk sectors (not Ext2 blocks) in use by this inode, not counting the actual inode structure nor directory entries linking to the inode.

flags: u32

Flags (see below)

os_specific1: u32

Operating System Specific value #1

dbp: [u32; 12]

Direct Block Pointers

sibp: u32

Singly Indirect Block Pointer (Points to a block that is a list of block pointers to data)

dibp: u32

Doubly Indirect Block Pointer (Points to a block that is a list of block pointers to Singly Indirect Blocks)

tibp: u32

Triply Indirect Block Pointer (Points to a block that is a list of block pointers to Doubly Indirect Blocks)

gen_no: u32

Generation number (Primarily used for NFS)

facl: u32

In Ext2 version 0, this field is reserved. In version >= 1, Extended attribute block (File ACL).

size_uh: u32

In Ext2 version 0, this field is reserved. In version >= 1, Upper 32 bits of file size (if feature bit set) if it’s a file, Directory ACL if it’s a directory

block_addr: u32

Block address of fragment

os_specific2: [u8; 12]

Operating System Specific Value #2

Implementations

impl Inode

pub fn new() -> Self

pub fn size(&self) -> u64

pub fn get_hardlinks(&self) -> u16

pub fn get_perms(&self) -> u16

pub fn is_dir(&self) -> bool

pub fn is_valid_block(block_no: u32) -> bool

fn get_blocks_no_from_u32_slice(slice: &[u32]) -> Vec<u32>

fn get_blocks_no_from_u8_slice(slice: &[u8]) -> Vec<u32>

pub fn get_blocks_no(&self) -> Vec<u32>

pub fn _get_sibp_blocks_no(sibp: u32, ext2: &Ext2) -> Vec<u32>

pub fn get_sibp_blocks_no(&self, ext2: &Ext2) -> Vec<u32>

pub fn _get_dibp_blocks_no(dibp: u32, ext2: &Ext2) -> Vec<u32>

pub fn get_dibp_blocks_no(&self, ext2: &Ext2) -> Vec<u32>

pub fn _get_tibp_blocks_no(tibp: u32, ext2: &Ext2) -> Vec<u32>

pub fn get_tibp_blocks_no(&self, ext2: &Ext2) -> Vec<u32>