Struct kernel::fs::ext2::Ext2

pub struct Ext2 {
    diskno: u8,
    sector_size: usize,
    pub sblock: BaseSuperblock,
}

Current read/write use entire block to perform operations In the filesystem created to test it this means we read/write 16 sectors for each operations This is pretty ineffective and will probably need optimisation in later version

Fields
diskno: u8sector_size: usizesblock: BaseSuperblock
Implementations

impl Ext2

pub fn new(diskno: u8) -> Result<Self, u8>

pub fn is_valid(&self) -> bool

pub fn inode_to_bgroup(&self, inode: u32) -> u32

Convert an inode number to it’s corresponding Group Number

Arguments

inode - Inode number to convert

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();
let groupno = ext2.inode_to_bgroup(45);

pub fn inode_to_block(&self, inode: u32) -> u32

Convert an inode number to it’s corresponding Block Number inside it’s group

Arguments

inode - Inode number to convert

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();
let blockno = ext2.inode_to_block(45);

pub fn inode_to_offset(&self, inode: u32) -> u32

Convert an inode number to it’s offset inside block

Arguments

inode - Inode number to convert

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();
let offset = ext2.inode_to_offset(45);

pub fn inode_size(&self) -> u16

Give inode size indicated by Superblock

pub fn read_block(&self, block_no: u32) -> Vec<u8>

Read an entire block from disk

fn write_block(&mut self, block_no: u32, block: &[u8])

fn write_slice(&mut self, block_no: u32, offset: usize, slice: &[u8])

fn get_inode(&self, inodeno: usize) -> Inode

fn write_inode(&mut self, inodeno: usize, inode: &Inode)

fn get_gdt_entry(&self, entry: usize) -> GdtEntry

Read disk to recover Group Descriptor Table entry given an index

pub fn read_inode_map(&self, group: usize) -> Bitmap

pub fn read_block_map(&self, group: usize) -> Bitmap

pub fn write_block_map(&mut self, group: usize, map: Bitmap)

pub fn write_inode_map(&mut self, group: usize, map: Bitmap)

pub fn alloc_node(&mut self, group: usize) -> usize

pub fn free_node(&mut self, group: usize, inodeno: usize)

pub fn alloc_block(&mut self, group: usize) -> usize

pub fn get_inode_entry(&self, entry: usize) -> Inode

Read disk to recover inode struct correcponding to the index passed as parameter

Arguments

entry - Inode index to read

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();

// Inode 2 is always inode to root dir
let inode = ext2.get_inode_entry(2);
crate::kprintln!("{:#?}", inode);

pub fn dentry_find(&self, inodeno: usize, filename: &str) -> Option<Dentry>

Find file inside dentry given the dentry inode and file searched. Currently ignore error cases and remain basic

Arguments

inodeno - Inode number of a directory

filename - Filename to look for inside this directory

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();
// Look for "dev" inside inode 2 (inode 2 is the inode for root directory)
let dentry = ext2.dentry_find(2, "dev");
match dentry {
    None => crate::kprintln!("File not found"),
    Some(dir) => crate::kprintln!("Found: {:#?}", dir)
};

pub fn get_inode_of(&self, path: &str) -> Option<(usize, Inode)>

Find file given it’s path, start search from root directory

Arguments

path - Absolute path to searched entry

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();
let opt = ext2.get_inode_of("/dev/vga");
match opt {
    Some((inodeno, inode)) => crate::kprintln!("Found at inode {}:\n{:#?}", inodeno, inode);
    None => crate::kprintln!("Not found")
};

pub fn recurs_find(&self, path: &str, inodeno: usize) -> Option<(usize, Inode)>

Perform recursive call to find file pass as argument starting at inodeno

Arguments

path - Relative path to file from given inode directory number

inodeno - Inode number of directory

Example

let ext2 = Ext2::new(unsafe { DISKNO as u8 }).unwrap();
// Look for vga named entry inside directory represented by inode 13
let opt = ext2.recurs_find("vga", 13)
match opt {
    None => crate::kprintln!("Entry does not exist"),
    Some((inodeno, _)) => crate::kprintln!("Entry exist at inode {}", inodeno)
};

pub fn write_dentries( &mut self, block_no: u32, dentries: Vec<Dentry> ) -> Result<(), ()>

pub fn get_dentries(&self, block_no: u32) -> Vec<Dentry>

pub fn remove_dentry(&mut self, parent_inodeno: usize, dentry: Dentry)

pub fn add_dentry(&mut self, inodeno: usize, dentry: Dentry)