kernel::fs

Module mapping

Structs

FileSystemMapping 🔒

MappingNode

Enums

MappingError

Statics

FILESYSTEM_MAPPING 🔒

Functions

get_mapping

Retrieves the mapping for the given path.

mount

Mounts a given filesystem at the specified path.

on_all_matching_mappings

Traverses the filesystem mapping tree and applies a handler function to all matching mappings.

unmount_all

Unmounts all filesystems from the virtual filesystem. This function removes all mounted filesystems from the virtual filesystem, effectively clearing the filesystem mapping tree.