Struct kernel::fs::path::PathBuf

pub struct PathBuf {
    inner: String,
}

An owned, mutable path (akin to String).

This type provides methods like PathBuf::push that mutate the path in place. It also implements Deref to Path, meaning that all methods on Path slices are available on PathBuf values as well.

More details about the overall approach can be found in the crate documentation.

Examples

You can use PathBuf::push to build up a PathBuf from components:

let mut path = PathBuf::new();

path.push("/");
path.push("feel");
path.push("the");

However, PathBuf::push is best used for dynamic situations. This is a better way to do this when you know all of the components ahead of time:

let path: PathBuf = ["/", "feel", "the.force"].iter().collect();

We can still do better than this! Since these are all strings, we can use From::from:

let path = PathBuf::from(r"/feel/the.force");

Which method works best depends on what kind of situation you’re in.

Fields
inner: String
Implementations

impl PathBuf

pub fn new() -> PathBuf

Allocates an empty PathBuf.

Examples

let path = PathBuf::new();

pub fn with_capacity(capacity: usize) -> PathBuf

Creates a new PathBuf with a given capacity used to create the internal String. See String::with_capacity.

Examples

let mut path = PathBuf::with_capacity(10);
let capacity = path.capacity();

// This push is done without reallocating
path.push("/");

assert_eq!(capacity, path.capacity());

pub fn as_path(&self) -> &Path

Coerces to a Path slice.

Examples

let p = PathBuf::from("/test");
assert_eq!(Path::new("/test"), p.as_path());

pub fn push<P: AsRef<Path>>(&mut self, path: P)

Extends self with path.

If path is absolute, it replaces the current path.

Examples

Pushing a relative path extends the existing path:

let mut path = PathBuf::from("/tmp");
path.push("file.bk");
assert_eq!(path, PathBuf::from("/tmp/file.bk"));

Pushing an absolute path replaces the existing path:

let mut path = PathBuf::from("/tmp");
path.push("/etc");
assert_eq!(path, PathBuf::from("/etc"));

fn _push(&mut self, path: &Path)

pub fn pop(&mut self) -> bool

Truncates self to self.parent.

Returns false and does nothing if self.parent is None. Otherwise, returns true.

Examples

let mut p = PathBuf::from("/test/test.rs");

p.pop();
assert_eq!(Path::new("/test"), p);
p.pop();
assert_eq!(Path::new("/"), p);

pub fn set_file_name<S: AsRef<str>>(&mut self, file_name: S)

Updates self.file_name to file_name.

If self.file_name was None, this is equivalent to pushing file_name.

Otherwise it is equivalent to calling pop and then pushing file_name. The new path will be a sibling of the original path. (That is, it will have the same parent.)

Examples

let mut buf = PathBuf::from("/");
assert!(buf.file_name() == None);
buf.set_file_name("bar");
assert!(buf == PathBuf::from("/bar"));
assert!(buf.file_name().is_some());
buf.set_file_name("baz.txt");
assert!(buf == PathBuf::from("/baz.txt"));

fn _set_file_name(&mut self, file_name: &str)

pub fn into_string(self) -> String

Consumes the PathBuf, yielding its internal String storage.

Examples

let p = PathBuf::from("/the/head");
let bytes = p.into_string();

pub fn into_boxed_path(self) -> Box<Path>

Converts this PathBuf into a boxed Path.

pub fn capacity(&self) -> usize

Invokes capacity on the underlying instance of String.

pub fn clear(&mut self)

Invokes clear on the underlying instance of String.

pub fn reserve(&mut self, additional: usize)

Invokes reserve on the underlying instance of String.

pub fn reserve_exact(&mut self, additional: usize)

Invokes reserve_exact on the underlying instance of String.

pub fn shrink_to_fit(&mut self)

Invokes shrink_to_fit on the underlying instance of String.

pub fn shrink_to(&mut self, min_capacity: usize)

Invokes shrink_to on the underlying instance of String.