kernel::fs

Struct File

pub struct File {
    filesystem: Arc<dyn FileSystem>,
    path: Box<Path>,
    inode: FileNode,
    position: u64,
    is_terminal: bool,
    blocking_mode: BlockingMode,
    access_helper: AccessHelper,
    file_access: FileAccess,
}

A handle to a file, it has the inode which controls the properties of the node in the filesystem

Fields
filesystem: Arc<dyn FileSystem>path: Box<Path>inode: FileNodeposition: u64is_terminal: boolblocking_mode: BlockingModeaccess_helper: AccessHelperfile_access: FileAccess
Implementations

impl File

pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, FileSystemError>

pub fn open_blocking<P: AsRef<Path>>( path: P, blocking_mode: BlockingMode, open_options: OpenOptions, ) -> Result<Self, FileSystemError>

pub fn from_inode<P: AsRef<Path>>( inode: FileNode, path: P, filesystem: Arc<dyn FileSystem>, position: u64, blocking_mode: BlockingMode, file_access: FileAccess, ) -> Result<Self, FileSystemError>

pub fn read(&mut self, buf: &mut [u8]) -> Result<u64, FileSystemError>

pub fn write(&mut self, buf: &[u8]) -> Result<u64, FileSystemError>

pub fn flush(&mut self) -> Result<(), FileSystemError>

pub fn seek(&mut self, position: u64) -> Result<(), FileSystemError>

pub fn filesize(&self) -> u64

pub fn path(&self) -> &Path

pub fn read_to_end(&mut self) -> Result<Vec<u8>, FileSystemError>

pub fn is_blocking(&self) -> bool

pub fn blocking_mode(&self) -> BlockingMode

pub fn set_blocking(&mut self, blocking_mode: BlockingMode)

pub fn is_terminal(&self) -> bool

pub fn set_terminal(&mut self, is_terminal: bool)

pub fn size(&self) -> u64

pub fn current_position(&self) -> u64

pub fn set_size(&mut self, size: u64) -> Result<(), FileSystemError>

pub fn clone_inherit(&self) -> Self

This is a move verbose method than Clone::clone, as I want it to be more explicit to the user that this is not a normal clone operation.