Trait kernel::fs::FsDriver

pub trait FsDriver {
    // Required methods
    fn device_name(&self) -> String;
    fn description(&self) -> String;
    fn root_dir(&self) -> INodeType;
    fn read_dir(&self, dir: INodeType, index: usize) -> Option<FsRawDirEntry>;
    fn lookup(&self, dir: INodeType, name: &str) -> Result<INodeType>;
    fn open(self: Arc<Self>, inode: INodeType) -> Result<Arc<dyn FsAccessToken>>;
    fn stat(&self, inode: INodeType) -> Option<FsRawMetaData>;

    // Provided methods
    fn creat(
        self: Arc<Self>,
        _dir: INodeType,
        _name: &str
    ) -> Result<Arc<dyn FsAccessToken>> { ... }
    fn mkdir(self: Arc<Self>, _dir: INodeType, _name: &str) -> Result<()> { ... }
    fn rename(
        &self,
        _old_dir: INodeType,
        _old_name: &str,
        _new_dir: INodeType,
        _new_name: &str,
        _replace: bool
    ) -> Result<()> { ... }
    fn unlink(&self, _dir: INodeType, _name: &str) -> Result<()> { ... }
}

Required Methods

fn device_name(&self) -> String

Device name if mounted on physical device, otherwise name of file system driver

fn description(&self) -> String

Return mount options as string

fn root_dir(&self) -> INodeType

Returns the inode of the root directory

fn read_dir(&self, dir: INodeType, index: usize) -> Option<FsRawDirEntry>

Reads the specified directory

fn lookup(&self, dir: INodeType, name: &str) -> Result<INodeType>

Searches for a file system object with the specified file name in the specified directory.

fn open(self: Arc<Self>, inode: INodeType) -> Result<Arc<dyn FsAccessToken>>

Opens a file with the specified inode

fn stat(&self, inode: INodeType) -> Option<FsRawMetaData>

Obtains metadata for the specified inode

Provided Methods

fn creat( self: Arc<Self>, _dir: INodeType, _name: &str ) -> Result<Arc<dyn FsAccessToken>>

fn mkdir(self: Arc<Self>, _dir: INodeType, _name: &str) -> Result<()>

fn rename( &self, _old_dir: INodeType, _old_name: &str, _new_dir: INodeType, _new_name: &str, _replace: bool ) -> Result<()>

fn unlink(&self, _dir: INodeType, _name: &str) -> Result<()>