Trait kernel::fs::Type

pub trait Type {
    type Context: Context<Self> + ?Sized;
    type Data: PointerWrapper + Send + Sync = ();

    const SUPER_TYPE: Super;
    const NAME: &'static CStr;
    const FLAGS: i32;

    fn fill_super(
        data: <Self::Context as Context<Self>>::Data, 
        sb: NewSuperBlock<'_, Self>
    ) -> Result<&SuperBlock<Self>>;
}

A file system type.

Required Associated Types

type Context: Context<Self> + ?Sized

The context used to build fs configuration before it is mounted or reconfigured.

Provided Associated Types

type Data: PointerWrapper + Send + Sync = ()

Data associated with each file system instance.

Required Associated Constants

const SUPER_TYPE: Super

Determines how superblocks for this file system type are keyed.

const NAME: &'static CStr

The name of the file system type.

const FLAGS: i32

The flags of this file system type.

It is a combination of the flags in the flags module.

Required Methods

fn fill_super(
data: <Self::Context as Context<Self>>::Data,
sb: NewSuperBlock<'_, Self>
) -> Result<&SuperBlock<Self>>

Initialises a super block for this file system type.