kernel::graphics::vga

Struct VgaDisplay

pub struct VgaDisplay {
    fb_info: FrameBufferInfo,
    memory: VirtualSpace<[u8]>,
}

Fields
fb_info: FrameBufferInfomemory: VirtualSpace<[u8]>
Implementations

impl VgaDisplay

fn new(framebuffer: Framebuffer) -> Self

pub fn put_pixel(&mut self, x: usize, y: usize, color: Pixel)

pub fn clear(&mut self)

pub fn blit_inner_ranges( &mut self, src: (usize, usize), dest: (usize, usize), width: usize, height: usize, )

pub fn blit( &mut self, src_buffer: &[u8], src_framebuffer_info: &FrameBufferInfo, src: (usize, usize), dest: (usize, usize), width: usize, height: usize, )

unsafe fn blit_fast( &mut self, src_buffer: &[u8], src_framebuffer_info: &FrameBufferInfo, src: (usize, usize), dest: (usize, usize), width: usize, height: usize, )

blit the src framebuffer to the current framebuffer fast here means that we assume the src and dest have the same format

fn blit_slow( &mut self, src_buffer: &[u8], src_framebuffer_info: &FrameBufferInfo, src: (usize, usize), dest: (usize, usize), width: usize, height: usize, )

pub fn clear_rect( &mut self, dest_x: usize, dest_y: usize, width: usize, height: usize, color: Pixel, )