Struct kernel::hal::PhysicalAddress

pub struct PhysicalAddress(/* private fields */);

Implementations

impl PhysicalAddress

pub const fn direct_map<T>(&self) -> *mut T

Gets the pointer corresponding to the specified physical address.

pub const fn direct_unmap(va: usize) -> PhysicalAddress

impl PhysicalAddress

pub const NULL: Self = _

pub const fn new(val: PhysicalAddressRepr) -> Self

pub const fn from_usize(val: usize) -> Self

pub const fn from_u64(val: u64) -> Self

pub const fn as_repr(&self) -> PhysicalAddressRepr

pub const fn as_u64(&self) -> u64

pub const fn as_usize(&self) -> usize

pub const unsafe fn identity_map<T>(&self) -> *mut T

Gets a pointer identical to the specified physical address.

Safety

Pointers of this form may not map to some memory.