Function kernel::heap_allocator::split_hole

fn split_hole(hole: HoleInfo, required_layout: Layout) -> Option<Allocation>

Split the given hole into (front_padding,hole,back_padding) if it’s big enough to hold the given layout with the required alignment. - Front padding occurs when the required alignment is higher than that of the hole. - Back padding occurs when the layout’s size is smaller than the hole.