Struct kernel::heap_allocator::HoleList

pub struct HoleList {
    first: Hole,
}

Fields
first: Hole
Implementations

impl HoleList

pub unsafe fn new(hole_addr: VAddr, hole_size: usize) -> HoleList

Create a new HoleList that contains the given hole. Unsafe because it is undefined bahaviour if the address passes is invalid or if [hole_addr, hole_addr+size) is used somewhere.

pub fn allocate_first_fit( &mut self, layout: Layout ) -> Result<*mut u8, AllocError>

Search for a big enough hole for the given Layout with its required alignment. This uses the first-fit strategy, and so is O(n)

pub unsafe fn free(&mut self, ptr: *mut u8, layout: Layout)

Free an allocation defined by ptr and layout. Unsafe because if ptr was not returned by a call to allocate_first_fit, undefined behaviour may occur. Deallocates by walking the list and inserts the given hole at the correct position. If the freed block is adjacent to another one, they are merged.

pub fn get_min_size() -> usize