Module kernel::hil

Public traits for interfaces between Tock components.

Modules

adc
Interfaces for analog to digital converter peripherals.

crc
Interface for CRC computation.

dac
Interface for digital to analog converters.

flash
Interface for reading, writing, and erasing flash storage pages.

gpio
Interface for direct control of GPIO pins.

gpio_async
Interface for GPIO pins that require split-phase operation to control.

i2c
Interface for I2C master and slave peripherals.

led
Interface for LEDs that abstract away polarity and pin.

ninedof
Interface for chips that provide 9DOF functionality.

nonvolatile_storage
Generic interface for nonvolatile memory.

radio
Interface for sending and receiving IEEE 802.15.4 packets.

rng
Interfaces for accessing a random number generator.

spi
Interfaces for SPI master and slave communication.

symmetric_encryption
Interfaces for accessing encryption and decryption of symmetric ciphers.

temperature
Interface for sampling a temperature sensor.

time
Hardware agnostic interfaces for counter-like resources.

uart
Interfaces for UART communications.

watchdog
Interface for a watchdog timer.

Traits

Controller
Shared interface for configuring components.