Trait kernel::hil::gpio::Pin

pub trait Pin {
    fn make_output(&self);
    fn make_input(&self);
    fn disable(&self);
    fn set(&self);
    fn clear(&self);
    fn toggle(&self);
    fn read(&self) -> bool;
    fn enable_interrupt(&self, identifier: usize, mode: InterruptMode);
    fn disable_interrupt(&self);
}

Interface for synchronous GPIO pins.

Required Methods

fn make_output(&self)

Configure the GPIO pin as an output pin.

fn make_input(&self)

Configure the GPIO pin as an input pin.

fn disable(&self)

Disable the GPIO pin and put it into its lowest power mode.

fn set(&self)

Set the GPIO pin high. It must be an output.

fn clear(&self)

Set the GPIO pin low. It must be an output.

fn toggle(&self)

Toggle the GPIO pin. It must be an output.

fn read(&self) -> bool

Get the current state of an input GPIO pin.

fn enable_interrupt(&self, identifier: usize, mode: InterruptMode)

Enable an interrupt on the GPIO pin. It must be configured as an interrupt. The identifier can be any value and will be returned to you when the interrupt on this pin fires.

fn disable_interrupt(&self)

Disable the interrupt for the GPIO pin.