Module kernel::hil::radio

Interface for sending and receiving IEEE 802.15.4 packets.

Hardware independent interface for an 802.15.4 radio. Note that configuration commands are asynchronous and must be committed with a call to config_commit. For example, calling set_address will change the source address of packets but does not change the address stored in hardware used for address recognition. This must be committed to hardware with a call to config_commit. Please see the relevant TRD for more details.

Reexports

use returncode::ReturnCode;

Structs

Header

Enums

RadioMacLen

Constants

MAX_BUF_SIZE

MAX_FRAME_SIZE

MAX_MTU

MFR_SIZE

MIN_FRAME_SIZE

MIN_MHR_SIZE
These constants are used for interacting with the SPI buffer, which contains a 1-byte SPI command, a 1-byte PHY header, and then the 802.15.4 frame. In theory, the number of extra bytes in front of the frame can depend on the particular method used to communicate with the radio, but we leave this as a constant in this generic trait for now.

MIN_PAYLOAD_OFFSET

PSDU_OFFSET

Traits

ConfigClient

PowerClient

Radio

RadioConfig
Configure the 802.15.4 radio.

RadioCrypto

RadioData

RxClient

TxClient