Module kernel::hil::spi

Interfaces for SPI master and slave communication.

Reexports

use core::option::Option;

use returncode::ReturnCode;

Enums

ClockPhase
Which clock edge values are sampled on

ClockPolarity
Values for the clock polarity (idle state or CPOL)

DataOrder
Values for the ordering of bits

Traits

SpiMaster
The SpiMaster trait for interacting with SPI slave devices at a byte or buffer level.

SpiMasterClient

SpiMasterDevice
SPIMasterDevice provides a chip-specific interface to the SPI Master hardware. The interface wraps the chip select line so that chip drivers cannot communicate with different SPI devices.

SpiSlave

SpiSlaveClient

SpiSlaveDevice
SPISlaveDevice provides a chip-specific interface to the SPI Slave hardware. The interface wraps the chip select line so that chip drivers cannot communicate with different SPI devices.