Module kernel::hil::symmetric_encryption

Interfaces for accessing encryption and decryption of symmetric ciphers.

Only AES-128-ctr supported at the moment.

The interface is supposed to work for hardware supported crypto but should work for software implemented crypto as well.

State Machine:

init()

set_key()

aes128_crypt_ctr(): can be used arbitrary number of times.

Reexports

use returncode::ReturnCode;

Traits

Client

SymmetricEncryption