Trait kernel::hil::symmetric_encryption::SymmetricEncryption

pub trait SymmetricEncryption {
    fn set_client(&self, client: &'static Client);
    fn init(&self);
    fn set_key(&self, key: &'static mut [u8], len: usize) -> &'static mut [u8];
    fn aes128_crypt_ctr(
        &self, 
        data: &'static mut [u8], 
        init_ctr: &'static mut [u8], 
        len: usize
    );
}

Required Methods

fn set_client(&self, client: &'static Client)

Set the callback client.

fn init(&self)

Initialization of the chip register

fn set_key(&self, key: &'static mut [u8], len: usize) -> &'static mut [u8]

Configure encryption/decryption key assumes that key size is 16, 24 or 32 bytes

fn aes128_crypt_ctr(
&self,
data: &'static mut [u8],
init_ctr: &'static mut [u8],
len: usize
)

encryption and decryption for aes in counter mode because only the encryption-mode of the cipher only one method is needed other chips perhaps only ignore "init_ctr" and assume all is performed in HW