Trait kernel::hil::watchdog::Watchdog

pub trait Watchdog {
    fn start(&self, period: usize);
    fn stop(&self);
    fn tickle(&self);
}

Required Methods

fn start(&self, period: usize)

Enable the watchdog timer. Period is the time in milliseconds the watchdog will timeout if not serviced.

fn stop(&self)

Disable the watchdog timer.

fn tickle(&self)

Service the watchdog to let the hardware know the application is still executing.