Function kernel::init::common::ffi_init3

pub unsafe fn ffi_init3<T, A1, A2, A3>(
    init_func: unsafe extern "C" fn(_: *mut T, arg1: A1, arg2: A2, arg3: A3), 
    arg1: A1, 
    arg2: A2, 
    arg3: A3
) -> impl PinInit<Opaque<T>>

Create an initializer using the given initializer function from C.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.