Module kernel::irq

Interrupts and interrupt chips.

See https://www.kernel.org/doc/Documentation/core-api/genericirq.rst.

C headers: include/linux/irq.h and include/linux/interrupt.h.

Modules

flags

Container for interrupt flags.

Structs

ChainedGuard

A guard to call chained_irq_exit after chained_irq_enter was called.

Descriptor

Wraps the kernel’s struct irq_desc.

Domain

Wraps the kernel’s struct irq_domain.

IrqData

Wraps the kernel’s struct irq_data.

LockedIrqData

Wraps the kernel’s struct irq_data when it is locked.

Registration

The registration of an interrupt handler.

ThreadedRegistration

The registration of a threaded interrupt handler.

Type

Contains constants that describes how an interrupt can be triggered.

Enums

ExtraResult

Extra information returned by some of the Chip methods on success.

Return

The return value from interrupt handlers.

Traits

Chip

An irq chip.

FlowHandler

A high-level irq flow handler.

Handler

An irq handler.

ThreadedHandler

A threaded irq handler.

Functions

set_wake

Enables or disables power-management wake-on for the given irq number.

Type Definitions

HwNumber

The type of irq hardware numbers.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.