Module kernel::kasync::executor

Kernel support for executing futures.

Modules

workqueue

Kernel support for executing futures in C workqueues (struct workqueue_struct).

Structs

AutoStopHandle

A handle to an executor that automatically stops it on drop.

Traits

Executor

An environment for executing tasks.

RefWake

A waker that is wrapped in Ref for its reference counting.

Task

A task spawned in an executor.

Functions

ref_waker

Creates a Waker from a Ref<T>, where T implements the RefWake trait.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.