Trait kernel::kasync::executor::ArcWake

pub trait ArcWake: Send + Sync {
    fn wake_by_ref(self: ArcBorrow<'_, Self>);

    fn wake(self: Arc<Self>) { ... }
}

A waker that is wrapped in Arc for its reference counting.

Types that implement this trait can get a Waker by calling ref_waker.

Required Methods

fn wake_by_ref(self: ArcBorrow<'_, Self>)

Wakes a task up.

Provided Methods

fn wake(self: Arc<Self>)

Wakes a task up and consumes a reference.