Module kernel::linked_list

Linked lists.

TODO: This module is a work in progress.

Structs

Cursor

A list cursor that allows traversing a linked list and inspecting elements.

CursorMut

A list cursor that allows traversing a linked list and inspecting & mutating elements.

Links

The links used to link an object on a linked list.

List

A linked list.

Traits

GetLinks

A descriptor of list elements.

GetLinksWrapped

A descriptor of wrapped list elements.

Wrapper

Wraps an object to be inserted in a linked list.