Trait kernel::linked_list::GetLinks

pub trait GetLinks {
    type EntryType: ?Sized;
    fn get_links(data: &Self::EntryType) -> &Links<Self::EntryType>;
}

A descriptor of list elements.

It describes the type of list elements and provides a function to determine how to get the links to be used on a list.

A type that may be in multiple lists simultaneously neneds to implement one of these for each simultaneous list.

Associated Types

type EntryType: ?Sized

The type of the entries in the list.

Loading content...
Required methods

fn get_links(data: &Self::EntryType) -> &Links<Self::EntryType>

Returns the links to be used when linking an entry within a list.

Loading content...