Module kernel::mem

Allocator implementation

Structs

Allocator
Frame allocator gives out one or more pages.

OsAllocator

Constants

MAX_PAGE
Maximum number of pages. As QEMU and linker script kernel.ld are set to have 128MB of RAM, maximum number of pages can be calculated.

Statics

GA

__ALLOC

Functions

ALLOC

align_val
Align an address to upper bound according to specified order.

align_val_down
Align an address to lower bound according to specified order.

alloc_error

alloc_stack

debug

hartinit

init⚠
Initialize allocator and kernel page table This function should only be called in boot hart

page_down
Align an address to the begin of a page.

zero_volatile⚠