Struct kernel::mem::address::physical_address::PhysicalAddress

pub struct PhysicalAddress {
    bits: usize,
}

The PhysicalAddress struct represents a 56-bit physical address defined in the Sv39 page table format.

Fields
bits: usize
Implementations

impl PhysicalAddress

pub fn floor(&self) -> FrameNumber

Returns the FrameNumber that represents the frame that contains the physical address.

pub fn ceil(&self) -> FrameNumber

Returns the FrameNumber that represents the frame that contains the physical address, rounding up to the next frame if the physical address is not aligned to a frame.

pub fn page_offset(&self) -> usize

Returns the byte offset of the physical address within its containing frame.

pub fn is_aligned(&self) -> bool

Returns true if the physical address is aligned to a frame.

pub fn as_ptr(&self) -> *const u8

Returns a raw pointer to the physical address.

pub fn as_ptr_mut(&self) -> *mut u8

Returns a mutable raw pointer to the physical address.

pub fn as_ref<T>(&self) -> &'static T

pub fn as_mut<T>(&self) -> &'static mut T