Module kernel::mem::segment

Structs

InternalBitFlags

Iter

IterRaw

KERNEL_SPACE

MapPermission

PageSegment

The PageSegment struct represents a consecutive range of pages, which are mapped to frames in the same method (Identical or Framed) and have the same permissions.

PageSet

The PageSet struct represents a collection of related PageSegment.

Enums

MapType

Constants

NUM_FLAGS ğŸ”’

OPTIONS ğŸ”’

OPTIONS_NAMES ğŸ”’

Statics

LAZY ğŸ”’

Functions

__stability ğŸ”’

__static_ref_initialize ğŸ”’

bss_end ğŸ”’ âš

bss_stack_start ğŸ”’ âš

data_end ğŸ”’ âš

data_start ğŸ”’ âš

kernel_end ğŸ”’ âš

rodata_end ğŸ”’ âš

rodata_start ğŸ”’ âš

text_end ğŸ”’ âš

text_start ğŸ”’ âš

trampoline_start ğŸ”’ âš