Struct kernel::memory::buddy_allocator::BuddyAllocator

pub struct BuddyAllocator {
    bins: [BTreeSet<PhysicalAddress>; 11],
}

Fields
bins: [BTreeSet<PhysicalAddress>; 11]

The bins of free blocks, where bin i contains blocks of size 2^i. Uses BTreeSet to store the blocks in each bin, for efficient buddy location. Each block is stored as the physical address of the start of the block. The actual frames can be constructed for each block using the start address and the order of the block.

Implementations

impl BuddyAllocator

pub fn new() -> BuddyAllocator

pub fn add_range(&mut self, range: Range<Frame>)

Add a range of Frames to this allocator, marking them free to allocate.

pub fn allocate_n(&mut self, num_bytes: usize) -> Option<PhysicalAddress>

Allocate (at least) n contiguous bytes from this allocator. Returns None if this allocator can't satisfy the allocation. The requested number of bytes must be a power-of-two.

pub fn free_n(&mut self, start: PhysicalAddress, num_bytes: usize)

Free the given block (starting at start and of size n frames). n must be a power-of-2.

fn allocate_block(&mut self, order: usize) -> Option<PhysicalAddress>

Tries to allocate a block of the given order. If no blocks of the correct size are available, tries to recursively split a larger block to form a block of the requested size.

fn free_block(&mut self, start: PhysicalAddress, order: usize)

Free a block starting at start of order order.

fn buddy_of(block_start: PhysicalAddress, order: usize) -> PhysicalAddress

Finds the starting frame of the block that is the buddy of the block of order order, starting at block_start.