kernel::memory_management

Module virtual_memory_mapper

This very specific to 64-bit x86 architecture, if this is to be ported to other architectures this will need to be changed

Modules

flags

Structs

PageDirectoryTable 🔒

PageDirectoryTablePtr 🔒

VirtualMemoryMapEntry

VirtualMemoryMapper

Constants

ADDR_MASK 🔒

KERNEL_L3_INDEX_END 🔒

KERNEL_L3_INDEX_START 🔒

KERNEL_L3_PROCESS_INDEX_END 🔒

KERNEL_L3_PROCESS_INDEX_START 🔒

KERNEL_L4_INDEX 🔒

KERNEL_PROCESS_VIRTUAL_ADDRESS_START

MAX_USER_VIRTUAL_ADDRESS

NUM_USER_L4_INDEXES 🔒

Statics

KERNEL_VIRTUAL_MEMORY_MANAGER 🔒

Functions

clone_current_vm_as_user

get_current_vm

get_l1 🔒

get_l2 🔒

get_l3 🔒

get_l4 🔒

init_kernel_vm

is_address_mapped_in_kernel

map_kernel

switch_to_kernel⚠

Safety

unmap_kernel

is_allocated is used to indicate if the physical pages were allocated by the caller i.e. when we called map_kernel, the physical_address is None and we will allocate the pages, and thus when calling this function, you should pass is_allocated = true