kernel::memory_management

Module virtual_space

Structs

VirtualSpace

A wrapper over memory that is defined by its physical address. We map this memory in virtual space, and return a pointer to it.

VirtualSpaceAllocator 🔒

VirtualSpaceEntry 🔒

Enums

VirtualSpaceError

Statics

VIRTUAL_SPACE_ALLOCATOR 🔒

Functions

allocate_and_map_virtual_space 🔒

deallocate_virtual_space 🔒

debug_blocks

Type Aliases

Result 🔒