Trait kernel::object::KernelObject

pub trait KernelObject {
    fn id(&self) -> KernelObjectId;
}

This trait should be implemented by all types that implement kernel objects, and allows common code to be generic over all kernel objects. Kernel objects are generally handled as Arc<T> where T is the type implementing KernelObject, and so interior mutability should be used for data that needs to be mutable within the kernel object.

Required methods

fn id(&self) -> KernelObjectId

Loading content...