Struct kernel::pci::ide::IDEController

pub struct IDEController {
    devices: [IDEDevice; 4],
}

Fields
devices: [IDEDevice; 4]
Implementations

impl IDEController

pub const fn new() -> Self

pub fn initialize( &mut self, bar0: u32, bar1: u32, bar2: u32, bar3: u32, bar4: u32 ) -> Result<(), u8>

fn wait_irq()

fn irq()

fn read(channel: &IDEChannelRegisters, reg: u8) -> u8

fn write(channel: &IDEChannelRegisters, reg: u8, data: u8)

fn read_buffer( channel: &IDEChannelRegisters, reg: u8, buffer: &mut [u32], quads: u32 )

fn polling(channel: &IDEChannelRegisters, advanced_check: u32) -> Result<(), u8>

fn print_error(&mut self, drive: u8, err: u8) -> u8

pub fn read_sectors( &mut self, drive: u8, numsects: u8, lba: u32, edi: u32 ) -> Result<(), u8>

Read sector from a drive

Parameters:

drive: drive number which can be from 0 to 3

numsects: number of sectors to be read. If 0, the ATA controller will now we want 256 sectors

lba: LBA address –> index of the sector, which allows us to acces disks up to 2TB

edi: adress of the buffer we want to fill

pub fn write_sectors( &mut self, drive: u8, numsects: u8, lba: u32, edi: u32 ) -> Result<(), u8>

Read sector from a drive

Parameters:

drive: drive number which can be from 0 to 3

numsects: number of sectors to write. If 0, the ATA controller will now we want 256 sectors

lba: LBA address –> index of the sector, which allows us to access disks up to 2TB

edi: adress of the buffer we want to fill