Trait kernel::pci::PciInterruptConfigurator

pub trait PciInterruptConfigurator {
    // Required methods
    fn configure_msi(
        &self,
        function: PciAddress,
        msi: &mut MsiCapability
    ) -> Arc<Event>;
    fn configure_msix(
        &self,
        function: PciAddress,
        table_bar: Bar,
        msix: &mut MsixCapability
    ) -> Arc<Event>;
}

Required Methods

fn configure_msi( &self, function: PciAddress, msi: &mut MsiCapability ) -> Arc<Event>

Create an Event that is signalled when an interrupt arrives from the specified PCI device. The device must support configuration of its interrupts via the passed MSI capability.

fn configure_msix( &self, function: PciAddress, table_bar: Bar, msix: &mut MsixCapability ) -> Arc<Event>

Create an Event that is signalled when an interrupt arrives from the specified PCI device. The device must support configuration of its interrupts via the passed MSI-X capability.