Trait kernel::platform::systick::SysTick

pub trait SysTick {
    fn set_timer(&self, us: u32);
    fn value(&self) -> u32;
    fn overflowed(&self) -> bool;
    fn reset(&self);
    fn enable(&self, with_interrupt: bool);
    fn overflow_fired() -> bool;
}

Interface for the system tick timer.

Required Methods

fn set_timer(&self, us: u32)

Sets the timer as close as possible to the given interval in microseconds. The clock is 24-bits wide and specific timing is dependent on the driving clock. Increments of 10ms are most accurate and, in practice 466ms is the approximate maximum.

fn value(&self) -> u32

Returns the time left in approximate microseconds

fn overflowed(&self) -> bool

fn reset(&self)

fn enable(&self, with_interrupt: bool)

fn overflow_fired() -> bool