Trait kernel::power::Operations

pub trait Operations {
    type Data: ForeignOwnable + Sync + Send;

    fn suspend(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result { ... }
    fn resume(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result { ... }
    fn freeze(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result { ... }
    fn restore(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result { ... }
}

Corresponds to the kernel’s struct dev_pm_ops.

It is meant to be implemented by drivers that support power-management operations.

Required Associated Types

type Data: ForeignOwnable + Sync + Send

The type of the context data stored by the driver on each device.

Provided Methods

fn suspend(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result

Called before the system goes into a sleep state.

fn resume(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result

Called after the system comes back from a sleep state.

fn freeze(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result

Called before creating a hibernation image.

fn restore(_data: <Self::Data as ForeignOwnable>::Borrowed<'_>) -> Result

Called after the system is restored from a hibernation image.