Module kernel::prelude

The kernel prelude.

These are most common items used by Rust code in the kernel, intended to be imported by all Rust code, for convenience.

Examples

use kernel::prelude::*;

Re-exports

pub use super::pr_alert;

pub use super::pr_cont;

pub use super::pr_crit;

pub use super::pr_emerg;

pub use super::pr_err;

pub use super::pr_info;

pub use super::pr_notice;

pub use super::pr_warn;

pub use super::static_assert;

pub use super::KernelModule;

Macros

module
Declares a kernel module.

Structs

String
A UTF-8–encoded, growable string.

Traits

ToOwned
A generalization of Clone to borrowed data.

Type Definitions

KernelResult
A Result with an Error error type.