Module kernel::process

Support for creating and running userspace applications.

Reexports

use callback::AppId;

use common::RingBuffer;

use common::Queue;

use common::VolatileCell;

use container;

use core::mem;

use core::ptr;

use core::slice;

use core::str;

use core::cell::Cell;

use core::fmt::Write;

use core::intrinsics;

use core::ptr::read_volatile;

use core::ptr::write_volatile;

use platform::mpu;

use returncode::ReturnCode;

use syscall::Syscall;

use common::math;

Structs

FunctionCall

LoadInfo

LoadResult

Process

StoredRegs

Enums

Error

FaultResponse

IPCType

State

Task

Statics

APP_FAULT

HAVE_WORK

PROCS

SCB_REGISTERS

SYSCALL_FIRED

Functions

ipsr_isr_number_to_str

load⚠
Loads the process into memory

load_processes⚠
Helper function to load processes from flash into an array of active processes. This is the default template for loading processes, but a board is able to create its own load_processes() function and use that instead.

parse_and_validate_load_info⚠
Converts a pointer to memory to a LoadInfo struct

processes_blocked

schedule

switch_to_user⚠