Function kernel::process::load_processes

pub unsafe fn load_processes(
    start_of_flash: *const u8, 
    app_memory: &mut [u8], 
    procs: &mut [Option<Process<'static>>], 
    fault_response: FaultResponse
)

Helper function to load processes from flash into an array of active processes. This is the default template for loading processes, but a board is able to create its own load_processes() function and use that instead.

Processes are found in flash starting from the given address and iterating through Tock Binary Format headers. Processes are given memory out of the app_memory buffer until either the memory is exhausted or the allocated number of processes are created, with process structures placed in the provided array. How process faults are handled by the kernel is also selected.