Struct kernel::process::LoadResult

struct LoadResult {
    init_fn: usize,
    app_heap_start: *const u8,
    stack_data_boundary: *const u8,
    fixed_len: u32,
    package_name: &'static str,
}

Fields
init_fn: usize

The absolute address of the process entry point (i.e. _start).

app_heap_start: *const u8

The lowest free address in process memory after allocating space for the stack, loading the GOT, data and BSS

stack_data_boundary: *const u8

The initial stack pointer

fixed_len: u32

The length of the fixed segment, including the stack, GOT, .data, / BSS, and any necessary alignment

package_name: &'static str

The process's package name (used for IPC)