kernel::process

Struct Process

pub struct Process {

    vm: VirtualMemoryMapper,
    context: ProcessContext,
    id: u64,
    parent_id: u64,
    open_filesystem_nodes: BTreeMap<usize, FilesystemNode>,
    file_index_allocator: GoingUpAllocator,
    argv: Vec<String>,
    file_path: PathBuf,
    current_dir: Directory,
    stack_ptr_end: usize,
    stack_size: usize,
    heap_start: usize,
    heap_size: usize,
    heap_max: usize,
    priority: PriorityLevel,
    exit_code: i32,
    children_exits: BTreeMap<u64, i32>,

}

Fields
vm: VirtualMemoryMappercontext: ProcessContextid: u64parent_id: u64open_filesystem_nodes: BTreeMap<usize, FilesystemNode>file_index_allocator: GoingUpAllocatorargv: Vec<String>file_path: PathBufcurrent_dir: Directorystack_ptr_end: usizestack_size: usizeheap_start: usizeheap_size: usizeheap_max: usizepriority: PriorityLevelexit_code: i32children_exits: BTreeMap<u64, i32>
Implementations

impl Process

pub fn allocate_process( parent_id: u64, elf: &Elf, file: &mut File, argv: Vec<String>, current_dir: Directory, ) -> Result<Self, ProcessError>

pub unsafe fn switch_to_this_vm(&mut self)

Safety

Check virtual_memory_mapper::VirtualMemoryMapper::switch_to_this for more info

pub fn id(&self) -> u64

pub fn parent_id(&self) -> u64

pub fn is_user_address_mapped(&self, address: usize) -> bool

pub fn finish_stdio(&mut self)

pub fn push_fs_node<F: Into<FilesystemNode>>(&mut self, file: F) -> usize

pub fn attach_fs_node_to_fd<F: Into<FilesystemNode>>( &mut self, fd: usize, file: F, ) -> bool

pub fn get_fs_node(&mut self, fd: usize) -> Option<&mut FilesystemNode>

pub fn take_fs_node(&mut self, fd: usize) -> Option<FilesystemNode>

pub fn put_fs_node(&mut self, fd: usize, file: FilesystemNode)

pub fn exit(&mut self, exit_code: i32)

Sets the exit_code and prepare to release the resources held by this process The scheduler will handle the state of the process

pub fn add_child_exit(&mut self, pid: u64, exit_code: i32)

pub fn get_child_exit(&mut self, pid: u64) -> Option<i32>

pub fn add_to_heap(&mut self, increment: isize) -> Option<usize>

Add/Remove to/from the heap and return the previous end of the heap before the change If this is an Add, it will return the address of the new block If this is a Remove, the result will generally be useless Use with 0 to get the current heap end

pub fn get_current_dir(&self) -> &Directory

pub fn set_current_dir(&mut self, current_dir: Directory)

pub fn get_priority(&self) -> PriorityLevel

pub fn set_priority(&mut self, priority: PriorityLevel)

pub fn file_path(&self) -> &Path

impl Process

fn prepare_stack( vm: &mut VirtualMemoryMapper, argv: &[String], rsp: u64, stack_top: u64, ) -> (u64, u64, u64)

fn write_process_meta( vm: &mut VirtualMemoryMapper, process_meta_addr: usize, process_meta: ProcessMetadata, )