kernel::process

Module syscalls

Constants

SYSCALLS ğŸ”’

Functions

check_ptr ğŸ”’

handle_syscall

path_to_proc_absolute_path ğŸ”’

Get the absolute path, if the path is relative, it will use the current process working directory to get the absolute path. If the path is absolute, it will return it as is.

ptr_as_mut ğŸ”’

ptr_as_ref ğŸ”’

sys_arg_to_file_mappings_array ğŸ”’

Allocates space fro the mapping and copies them

sys_arg_to_mut_slice ğŸ”’

sys_arg_to_path ğŸ”’

sys_arg_to_slice ğŸ”’

sys_arg_to_str ğŸ”’

sys_arg_to_str_array ğŸ”’

Allocates space for the strings and copies them

sys_blocking_mode ğŸ”’

sys_chdir ğŸ”’

sys_close ğŸ”’

sys_create_pipe ğŸ”’

sys_exit ğŸ”’

sys_get_cwd ğŸ”’

sys_get_file_meta ğŸ”’

sys_get_time ğŸ”’

sys_graphics ğŸ”’

sys_inc_heap ğŸ”’

sys_open ğŸ”’

sys_open_dir ğŸ”’

sys_priority ğŸ”’

Set and Get process priority TODO: implement security levels, as now we can change the priority of any process

sys_read ğŸ”’

sys_read_dir ğŸ”’

sys_seek ğŸ”’

sys_set_file_meta ğŸ”’

sys_sleep ğŸ”’

sys_spawn ğŸ”’

sys_stat ğŸ”’

sys_wait_pid ğŸ”’

sys_write ğŸ”’

Type Aliases

Syscall ğŸ”’