Module kernel::random

Random numbers.

C header: include/linux/random.h

Functions

add_randomness

Contributes the contents of a byte slice to the kernel’s entropy pool.

getrandom

Fills a byte slice with random bytes generated from the kernel’s CSPRNG.

getrandom_nonblock

Fills a byte slice with random bytes generated from the kernel’s CSPRNG.