Module kernel::revocable

Revocable objects.

The Revocable type wraps other types and allows access to them to be revoked. The existence of a RevocableGuard ensures that objects remain valid.

Structs

AsyncRevocable

An object that can become inaccessible at runtime.

AsyncRevocableGuard

A guard that allows access to a revocable object and keeps it alive.

Revocable

An object that can become inaccessible at runtime.

RevocableGuard

A guard that allows access to a revocable object and keeps it alive.