Module kernel::scheduler::pit

Constants

BASE_FREQUENCY

How often the PIT Chip oscillates per second

DIVISOR_MAX

DIVISOR_MIN

PIT_INTERRUPT

Statics

DIVISOR

The Divisor - The higher the value, the less the frequency, the more time between interrupts

TIME_SINCE_BOOT

Without the Volatile, the compiler may optimize sleep into an infinite loop

Functions

get_frequency

Get Frequency

msleep

Millisecond-Sleep

nanosleep

Nanosecond-Sleep

pit_interrupt_handler

set_divisor

Set Divisor

set_frequency

Set Frequency

sleep

Sleep Seconds

tick

Tick

usleep

Microsecond-Sleep