Struct kernel::scheduler::Scheduler

pub struct Scheduler {
    pub running_task: Option<WrappedKernelObject<Arch>>,
    ready_queue: VecDeque<WrappedKernelObject<Arch>>,
}

Fields
running_task: Option<WrappedKernelObject<Arch>>ready_queue: VecDeque<WrappedKernelObject<Arch>>

List of Tasks ready to be scheduled. Every kernel object in this list must be a Task. Backed by a VecDeque so we can rotate objects in the queue efficiently.

Methods

impl Scheduler

pub fn new() -> Scheduler

pub fn add_task(
&mut self,
task_object: WrappedKernelObject<Arch>
) -> Result<(), ScheduleError>

pub fn drop_to_userspace(&mut self, arch: &Arch) -> !

Performs the first transistion from the kernel into userspace. On some platforms, this has to be done in a different way to how we'd replace the currently running task if we'd yielded or pre-empted out of an existing userspace context, and so this is handled specially.

The scheduler will always drop into userspace into the first task added to the ready queue. By controlling which Task is added first, the ecosystem can be sure that the correct Task is run first (whether the userspace layers take advantage of this is up to them - it would be more reliable to not depend on one process starting first, but this is an option).

pub fn switch_to_next(&mut self)

Switch to the next scheduled task. This is called when a task yields, or when we pre-empt a task that is hogging CPU time. If there is nothing to schedule, this is free to idle the CPU (including managing power), or steal work from another scheduling unit.