Module kernel::security

Linux Security Modules (LSM).

C header: include/linux/security.h.

Functions

binder_set_context_mgr

Calls the security modules to determine if the given task can become the manager of a binder context.

binder_transaction

Calls the security modules to determine if binder transactions are allowed from task from to task to.

binder_transfer_binder

Calls the security modules to determine if task from is allowed to send binder objects (owned by itself or other processes) to task to through a binder transaction.

binder_transfer_file

Calls the security modules to determine if task from is allowed to send the given file to task to (which would get its own file descriptor) through a binder transaction.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.