Struct kernel::str::CStr

#[repr(transparent)]
pub struct CStr(_);

A string that is guaranteed to have exactly one NUL byte, which is at the end.

Used for interoperability with kernel APIs that take C strings.

Implementations

impl CStr

pub const fn len(&self) -> usize

Returns the length of this string excluding NUL.

pub const fn len_with_nul(&self) -> usize

Returns the length of this string with NUL.

pub const fn is_empty(&self) -> bool

Returns true if the string only includes NUL.

pub unsafe fn from_char_ptr<'a>(ptr: *const c_char) -> &'a Self

Wraps a raw C string pointer.

Safety

ptr must be a valid pointer to a NUL-terminated C string, and it must last at least 'a. When CStr is alive, the memory pointed by ptr must not be mutated.

pub const fn from_bytes_with_nul(
bytes: &[u8]
) -> Result<&Self, CStrConvertError>

Creates a CStr from a [u8].

The provided slice must be NUL-terminated, does not contain any interior NUL bytes.

pub const unsafe fn from_bytes_with_nul_unchecked(bytes: &[u8]) -> &CStr

Creates a CStr from a [u8] without performing any additional checks.

Safety

bytes must end with a NUL byte, and should only have a single NUL byte (or the string will be truncated).

pub const fn as_char_ptr(&self) -> *const c_char

Returns a C pointer to the string.

pub fn as_bytes(&self) -> &[u8]

Convert the string to a byte slice without the trailing 0 byte.

pub const fn as_bytes_with_nul(&self) -> &[u8]

Convert the string to a byte slice containing the trailing 0 byte.