Struct kernel::Error

pub struct Error(_);

Generic integer kernel error.

The kernel defines a set of integer generic error codes based on C and POSIX ones. These codes may have a more specific meaning in some contexts.

Invariants

The value is a valid errno (i.e. >= -MAX_ERRNO && < 0).

Implementations

impl Error

pub const EPERM: Self

Operation not permitted.

pub const ENOENT: Self

No such file or directory.

pub const ESRCH: Self

No such process.

pub const EINTR: Self

Interrupted system call.

pub const EIO: Self

I/O error.

pub const ENXIO: Self

No such device or address.

pub const E2BIG: Self

Argument list too long.

pub const ENOEXEC: Self

Exec format error.

pub const EBADF: Self

Bad file number.

pub const ECHILD: Self

Exec format error.

pub const EAGAIN: Self

Try again.

pub const ENOMEM: Self

Out of memory.

pub const EACCES: Self

Permission denied.

pub const EFAULT: Self

Bad address.

pub const ENOTBLK: Self

Block device required.

pub const EBUSY: Self

Device or resource busy.

pub const EEXIST: Self

File exists.

pub const EXDEV: Self

Cross-device link.

pub const ENODEV: Self

No such device.

pub const ENOTDIR: Self

Not a directory.

pub const EISDIR: Self

Is a directory.

pub const EINVAL: Self

Invalid argument.

pub const ENFILE: Self

File table overflow.

pub const EMFILE: Self

Too many open files.

pub const ENOTTY: Self

Not a typewriter.

pub const ETXTBSY: Self

Text file busy.

pub const EFBIG: Self

File too large.

pub const ENOSPC: Self

No space left on device.

pub const ESPIPE: Self

Illegal seek.

pub const EROFS: Self

Read-only file system.

pub const EMLINK: Self

Too many links.

pub const EPIPE: Self

Broken pipe.

pub const EDOM: Self

Math argument out of domain of func.

pub const ERANGE: Self

Math result not representable.

pub const EDEADLK: Self

Resource deadlock would occur

pub const ENAMETOOLONG: Self

File name too long

pub const ENOLCK: Self

No record locks available

pub const ENOSYS: Self

Invalid system call number.

This error code is special: arch syscall entry code will return Self::ENOSYS if users try to call a syscall that doesn’t exist. To keep failures of syscalls that really do exist distinguishable from failures due to attempts to use a nonexistent syscall, syscall implementations should refrain from returning Self::ENOSYS.

pub const ENOTEMPTY: Self

Directory not empty.

pub const ELOOP: Self

Too many symbolic links encountered.

pub const EWOULDBLOCK: Self

Operation would block.

pub const ENOMSG: Self

No message of desired type.

pub const EIDRM: Self

Identifier removed.

pub const ECHRNG: Self

Channel number out of range.

pub const EL2NSYNC: Self

Level 2 not synchronized.

pub const EL3HLT: Self

Level 3 halted.

pub const EL3RST: Self

Level 3 reset.

pub const ELNRNG: Self

Link number out of range.

pub const EUNATCH: Self

Protocol driver not attached.

pub const ENOCSI: Self

No CSI structure available.

pub const EL2HLT: Self

Level 2 halted.

pub const EBADE: Self

Invalid exchange.

pub const EBADR: Self

Invalid request descriptor.

pub const EXFULL: Self

Exchange full.

pub const ENOANO: Self

No anode.

pub const EBADRQC: Self

Invalid request code.

pub const EBADSLT: Self

Invalid slot.

pub const EDEADLOCK: Self

Resource deadlock would occur.

pub const EBFONT: Self

Bad font file format.

pub const ENOSTR: Self

Device not a stream.

pub const ENODATA: Self

No data available.

pub const ETIME: Self

Timer expired.

pub const ENOSR: Self

Out of streams resources.

pub const ENONET: Self

Machine is not on the network.

pub const ENOPKG: Self

Package not installed.

pub const EREMOTE: Self

Object is remote.

pub const ENOLINK: Self

Link has been severed.

pub const EADV: Self

Advertise error.

pub const ESRMNT: Self

Srmount error.

pub const ECOMM: Self

Communication error on send.

pub const EPROTO: Self

Protocol error.

pub const EMULTIHOP: Self

Multihop attempted.

pub const EDOTDOT: Self

RFS specific error.

pub const EBADMSG: Self

Not a data message.

pub const EOVERFLOW: Self

Value too large for defined data type.

pub const ENOTUNIQ: Self

Name not unique on network.

pub const EBADFD: Self

File descriptor in bad state.

pub const EREMCHG: Self

Remote address changed.

pub const ELIBACC: Self

Can not access a needed shared library.

pub const ELIBBAD: Self

Accessing a corrupted shared library.

pub const ELIBSCN: Self

.lib section in a.out corrupted.

pub const ELIBMAX: Self

Attempting to link in too many shared libraries.

pub const ELIBEXEC: Self

Cannot exec a shared library directly.

pub const EILSEQ: Self

Illegal byte sequence.

pub const ERESTART: Self

Interrupted system call should be restarted.

pub const ESTRPIPE: Self

Streams pipe error.

pub const EUSERS: Self

Too many users.

pub const ENOTSOCK: Self

Socket operation on non-socket.

pub const EDESTADDRREQ: Self

Destination address required.

pub const EMSGSIZE: Self

Message too long.

pub const EPROTOTYPE: Self

Protocol wrong type for socket.

pub const ENOPROTOOPT: Self

Protocol not available.

pub const EPROTONOSUPPORT: Self

Protocol not supported.

pub const ESOCKTNOSUPPORT: Self

Socket type not supported.

pub const EOPNOTSUPP: Self

Operation not supported on transport endpoint.

pub const EPFNOSUPPORT: Self

Protocol family not supported.

pub const EAFNOSUPPORT: Self

Address family not supported by protocol.

pub const EADDRINUSE: Self

Address already in use.

pub const EADDRNOTAVAIL: Self

Cannot assign requested address.

pub const ENETDOWN: Self

Network is down.

pub const ENETUNREACH: Self

Network is unreachable.

pub const ENETRESET: Self

Network dropped connection because of reset.

pub const ECONNABORTED: Self

Software caused connection abort.

pub const ECONNRESET: Self

Connection reset by peer.

pub const ENOBUFS: Self

No buffer space available.

pub const EISCONN: Self

Transport endpoint is already connected.

pub const ENOTCONN: Self

Transport endpoint is not connected.

pub const ESHUTDOWN: Self

Cannot send after transport endpoint shutdown.

pub const ETOOMANYREFS: Self

Too many references: cannot splice.

pub const ETIMEDOUT: Self

Connection timed out.

pub const ECONNREFUSED: Self

Connection refused.

pub const EHOSTDOWN: Self

Host is down.

pub const EHOSTUNREACH: Self

No route to host.

pub const EALREADY: Self

Operation already in progress.

pub const EINPROGRESS: Self

Operation now in progress.

pub const ESTALE: Self

Stale file handle.

pub const EUCLEAN: Self

Structure needs cleaning.

pub const ENOTNAM: Self

Not a XENIX named type file.

pub const ENAVAIL: Self

No XENIX semaphores available.

pub const EISNAM: Self

Is a named type file.

pub const EREMOTEIO: Self

Remote I/O error.

pub const EDQUOT: Self

Quota exceeded.

pub const ENOMEDIUM: Self

No medium found.

pub const EMEDIUMTYPE: Self

Wrong medium type.

pub const ECANCELED: Self

Operation Canceled.

pub const ENOKEY: Self

Required key not available.

pub const EKEYEXPIRED: Self

Key has expired.

pub const EKEYREVOKED: Self

Key has been revoked.

pub const EKEYREJECTED: Self

Key was rejected by service.

pub const EOWNERDEAD: Self

Owner died.

For robust mutexes.

pub const ENOTRECOVERABLE: Self

State not recoverable.

pub const ERFKILL: Self

Operation not possible due to RF-kill.

pub const EHWPOISON: Self

Memory page has hardware error.

pub const ERESTARTSYS: Self

Restart the system call.

pub const ENOTSUPP: Self

Operation is not supported.

pub fn to_kernel_errno(self) -> c_int

Returns the kernel error code.