Module kernel::symbols

Linker-script constants and RISC-V constants (This module may be viewed as memory layout definitions)

Re-exports

pub use gen::*;

Modules

gen
This module is automatically generated with symbols_gen.rs.py, which contains all linker script symbols in kernel.ld and a wrapper function to safely get them.

Constants

MAXVA
Maximum virtual address supported on Sv39

NCPUS
Maximum supported CPU on machine

NMAXPROCS
Maximum process on machine.

PAGE_ORDER
Page order

PAGE_SIZE
Page size

SCHEDULER_INTERVAL
Scheduler timer interrupt interval

TRAMPOLINE_START
Address to map kernel and user trampoline

TRAPFRAME_START
Address to map trapframe

Functions

bss_range⚠

kernelvec⚠
kernelvec function in kernelvec.S

print_map_symbols

timervec⚠
m_trap_vector function in trap.S

userret⚠
userret function in trampoline.S

uservec⚠
uservec function in trampoline.S