Module kernel::symbols::gen

This module is automatically generated with symbols_gen.rs.py, which contains all linker script symbols in kernel.ld and a wrapper function to safely get them.

Statics

__bss_end

__bss_start

__data_end

__data_start

__heap_size

__heap_start

__kernel_stack_end

__kernel_stack_start

__rodata_end

__rodata_start

__text_end

__text_start

__trampoline_text_start

Functions

BSS_END

BSS_START

DATA_END

DATA_START

HEAP_SIZE

HEAP_START

KERNEL_STACK_END

KERNEL_STACK_START

RODATA_END

RODATA_START

TEXT_END

TEXT_START

TRAMPOLINE_TEXT_START