Module kernel::sync

The sync module provides synchronization primitives for concurrent programming.

Re-exports

pub use event_bus::wait_for_event;

pub use event_bus::Event;

pub use event_bus::EventBus;

pub use mutex::Mutex;

pub use mutex::MutexGuard;

Modules

event_bus 🔒

The event_bus module provides an EventBus that supports publish-subscribe-style communication between different tasks.

mutex 🔒

The mutex module provides a mutual exclusion primitive useful for protecting shared data.