Struct kernel::sync::Mutex

pub struct Mutex<T: ?Sized> { /* fields omitted */ }

Exposes the kernel’s struct mutex. When multiple threads attempt to lock the same mutex, only one at a time is allowed to progress, the others will block (sleep) until the mutex is unlocked, at which point another thread will be allowed to wake up and make progress.

A Mutex must first be initialised with a call to Mutex::init_lock before it can be used. The mutex_init macro is provided to automatically assign a new lock class to a mutex instance.

Since it may block, Mutex needs to be used with care in atomic contexts.

Implementations

impl<T> Mutex<T>

pub const unsafe fn new(t: T) -> Self

Constructs a new mutex.

Safety

The caller must call Mutex::init_lock before using the mutex.

impl<T: ?Sized> Mutex<T>

pub fn lock(&self) -> Guard<'_, Self>

Locks the mutex and gives the caller access to the data protected by it. Only one thread at a time is allowed to access the protected data.