Struct kernel::sync::Ref

pub struct Ref<T: RefCounted + ?Sized> { /* fields omitted */ }

A reference-counted pointer to an instance of T.

The reference count is incremented when new instances of Ref are created, and decremented when they are dropped. When the count reaches zero, the underlying T is also dropped.

Invariants

The value stored in RefCounted::get_count corresponds to the number of instances of Ref that point to that instance of T.

Implementations

impl<T: RefCounted> Ref<T>

pub fn try_new(contents: T) -> KernelResult<Self>

Constructs a new reference counted instance of T.

impl<T: RefCounted + ?Sized> Ref<T>

pub fn new_from(obj: &T) -> Self

Creates a new reference-counted pointer to the given instance of T.

It works by incrementing the current reference count as part of constructing the new pointer.

pub fn get_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to T iff the reference count is one. Otherwise returns None.

pub fn ptr_eq(a: &Self, b: &Self) -> bool

Determines if two reference-counted pointers point to the same underlying instance of T.

pub fn into_raw(obj: Self) -> *const T

Deconstructs a Ref object into a raw pointer.

It can be reconstructed once via Ref::from_raw.

pub unsafe fn from_raw(ptr: *const T) -> Self

Recreates a Ref instance previously deconstructed via Ref::into_raw.

Safety

ptr must have been returned by a previous call to Ref::into_raw. Additionally, it can only be called once for each previous call to ``Ref::into_raw`.