Struct kernel::sync::RevocableMutex

pub struct RevocableMutex<T: ?Sized> { /* fields omitted */ }

A mutex whose contents can become inaccessible at runtime.

Once access is revoked and all concurrent users complete (i.e., all existing instances of RevocableMutexGuard are dropped), the wrapped object is also dropped.

Examples

struct Example {
    a: u32,
    b: u32,
}

fn add_two(v: &RevocableMutex<Example>) -> Option<u32> {
    let guard = v.try_lock()?;
    Some(guard.a + guard.b)
}

fn example() {
    // SAFETY: We call `revocable_mutex_init` immediately below.
    let mut v = unsafe { RevocableMutex::new(Example { a: 10, b: 20 }) };
    // SAFETY: We never move out of `v`.
    let pinned = unsafe { Pin::new_unchecked(&mut v) };
    revocable_mutex_init!(pinned, "example::v");
    assert_eq!(add_two(&v), Some(30));
    v.revoke();
    assert_eq!(add_two(&v), None);
}

Implementations

impl<T> RevocableMutex<T>

pub unsafe fn new(data: T) -> Self

Creates a new revocable instance of the given data.

Safety

The caller must call RevocableMutex::init before using the revocable mutex.

impl<T: ?Sized> RevocableMutex<T>

pub fn try_lock(&self) -> Option<RevocableMutexGuard<'_, T>>

Tries to lock (and access) the [revocable] wrapped object.

Returns None if the object has been revoked and is therefore no longer accessible.

Returns a guard that gives access to the object otherwise; the object is guaranteed to remain accessible while the guard is alive. Callers are allowed to sleep while holding on to the returned guard.

pub fn revoke(&self)

Revokes access to and drops the wrapped object.

Revocation and dropping happens after ongoing accessors complete.