Trait kernel::sync::LockIniter

pub trait LockIniter {
    fn init_lock(
        self: Pin<&mut Self>,
        name: &'static CStr,
        key: &'static LockClassKey
    );
}

A lock that can be initialised with a single lock class key.

Required Methods

fn init_lock(
self: Pin<&mut Self>,
name: &'static CStr,
key: &'static LockClassKey
)

Initialises the lock instance so that it can be safely used.