Module kernel::syscall

Module for processing syscall

All functions that begin with sys_ will extract parameters from trap frame, turn pointers into Rust primitives and call corresponding functions with these parameters in kernel code.

For specifications and how to do syscalls, refer to syscall module in user crate.

Re-exports

pub use gen::*;

Modules

file
File-related syscalls

gen
This module is automatically generated with syscall_gen.rs.py, which contains all syscall ID.

Functions

arg_fd
Get file corresponding to a file descriptor

arg_int
Get the posth argument as i32 from syscall

arg_ptr
Get the posth argument as a pointer from syscall, return kernel-space pointer (involve security issues!)

arg_ptr_mut
Get the posth argument as a pointer from syscall, return kernel-space mutable pointer (involve security issues!)

arg_uint
Get the posth argument as usize from syscall

argraw
Get the posth argument from syscall

sys_exec
exec syscall entry

sys_exit
exit syscall entry

sys_fork
fork syscall entry

syscall
Process all syscall