Function kernel::syscall::handle_syscall

pub fn handle_syscall<P>(
    number: usize,
    a: usize,
    b: usize,
    c: usize,
    d: usize,
    e: usize
) -> usizewhere
    P: Platform,

This is the architecture-independent syscall handler. It should be called by the handler that receives the syscall (each architecture is free to do this however it wishes). The only parameter that is guaranteed to be valid is number; the meaning of the rest may be undefined depending on how many parameters the specific system call takes.