Struct kernel::system::System

pub struct System { /* fields omitted */ }

Implementations

impl System

pub unsafe fn init(info: &BootInfo, f: fn()) -> !

Initialize the system

pub const fn name() -> &'static str

Returns the name of current system.

pub const fn short_name() -> &'static str

Returns abbreviated name of current system.

pub const fn version<'a>() -> &'a Version<'a>

Returns the version of current system.

pub fn boot_flags() -> BootFlags

pub fn system_time() -> SystemTime

Returns the current system time.

pub const fn is_multi_processor_capable_kernel() -> bool

Returns whether the kernel is multiprocessor-capable.

pub unsafe fn activate_cpu(new_cpu: Box<Cpu>)

Add SMP-initialized CPU cores to the list of enabled cores.

SAFETY: THREAD UNSAFE. Do not call this function except when initializing SMP.

pub unsafe fn current_processor<'a>() -> &'a Cpu

Returns an instance of the current processor.

pub fn cpu<'a>(index: usize) -> &'a Cpu

pub unsafe fn cpu_mut<'a>(index: usize) -> &'a mut Cpu

pub unsafe fn sort_cpus<F>(compare: F) where
F: FnMut(&Box<Cpu>, &Box<Cpu>) -> Ordering,

Sorts the list of CPUs by processor ID.

SAFETY: THREAD UNSAFE. Do not call this function except when initializing SMP.

pub fn acpi() -> &'static AcpiTables<MyAcpiHandler>

pub fn smbios<'a>() -> Option<&'a SmBios>

pub fn current_device<'a>() -> &'a DeviceInfo

pub fn acpi_platform() -> PlatformInfo

pub unsafe fn reset() -> !

pub unsafe fn shutdown() -> !

pub fn main_screen() -> Bitmap<'static>

Get main screen

pub fn em_console<'a>() -> &'a mut EmConsole

pub fn set_stdout(stdout: Box<dyn Tty>)

pub fn stdout<'a>() -> &'a mut dyn Tty