Struct kernel::task::scheduler::Scheduler

pub struct Scheduler { /* private fields */ }

System Scheduler

Implementations

impl Scheduler

pub unsafe fn start(f: fn(_: usize), args: usize) -> !

Start scheduler and sleep forever

pub unsafe fn init_second()

pub fn is_enabled() -> bool

Returns whether or not the thread scheduler is running.

pub const fn is_multi_processor_capable() -> bool

pub fn current_state() -> SchedulerState

Returns the current state of the scheduler.

pub fn freeze(force: bool)

All threads will stop.

pub fn current_pid() -> ProcessId

Get the current process running on the current processor

pub fn current_thread() -> Option<ThreadHandle>

Get the current thread running on the current processor

pub fn current_personality<'a>() -> Option<&'a mut PersonalityContext>

Get the personality instance associated with the current thread

pub unsafe fn reschedule()

Perform the preemption

pub fn sleep_thread()

pub fn usage_per_cpu() -> usize

pub fn usage_total() -> usize

pub fn spawn_async(task: impl Future<Output = ()> + 'static)

Spawning asynchronous tasks

pub fn spawn_task(task: Task)

pub fn perform_tasks() -> !

Performing Asynchronous Tasks

pub fn exit() -> !

pub fn get_idle_statistics(vec: &mut Vec<u32>)

pub fn print_statistics(sb: &mut impl Write)

pub fn get_thread_statistics(sb: &mut impl Write)