Struct kernel::task::Task

pub struct Task { /* fields omitted */ }

Wraps the kernel’s struct task_struct.

Invariants

The pointer Task::ptr is non-null and valid. Its reference count is also non-zero.

Examples

The following is an example of getting the PID of the current thread with zero additional cost when compared to the C version:

use kernel::task::Task;

Task::current().pid();

Getting the PID of the current process, also zero additional cost:

use kernel::task::Task;

Task::current().group_leader().pid();

Getting the current task and storing it in some struct. The reference count is automatically incremented when creating State and decremented when it is dropped:

use kernel::task::Task;

struct State {
    creator: Task,
    index: u32,
}

impl State {
    fn new() -> Self {
        Self {
            creator: Task::current().clone(),
            index: 0,
        }
    }
}

Implementations

impl Task

pub fn current<'a>() -> TaskRef<'a>

Returns a task reference for the currently executing task/thread.

pub fn group_leader(&self) -> TaskRef<'_>

Returns the group leader of the given task.

pub fn pid(&self) -> pid_t

Returns the PID of the given task.

pub fn signal_pending(&self) -> bool

Determines whether the given task has pending signals.