Trait kernel::Module

pub trait Module: Sized + Sync {
    fn init(name: &'static CStr, module: &'static ThisModule) -> Result<Self>;
}

The top level entrypoint to implementing a kernel module.

For any teardown or cleanup operations, your type may implement Drop.

Required Methods

fn init(name: &'static CStr, module: &'static ThisModule) -> Result<Self>

Called at module initialization time.

Use this method to perform whatever setup or registration your module should do.

Equivalent to the module_init macro in the C API.