Module kernel::types

Kernel types.

C header: include/linux/types.h

Structs

ARef

An owned reference to an always-reference-counted object.

Bit

A bitmask.

False

Represents the false value for types wth Bool bound.

Mode

Permissions.

Opaque

Stores an opaque value.

ScopeGuard

Runs a cleanup function/closure when dropped.

True

Represents the true value for types with Bool bound.

Enums

Either

A sum type that always holds either a value of type L or R.

Traits

AlwaysRefCounted

Types that are always reference counted.

Bool

A trait for boolean types.

ForeignOwnable

Used to transfer ownership to and from foreign (non-Rust) languages.

TrailingZeros

Implemented by integer types that allow counting the number of trailing zeroes.

Functions

bit

Creates a bit mask with a single bit set.

bits_iter

Returns an iterator over the set bits of value.