Struct kernel::types::Opaque

#[repr(transparent)]
pub struct Opaque<T>(_);

Stores an opaque value.

This is meant to be used with FFI objects that are never interpreted by Rust code.

Implementations

impl<T> Opaque<T>

pub const fn new(value: T) -> Self

Creates a new opaque value.

pub const fn uninit() -> Self

Creates an uninitialised value.

pub fn get(&self) -> *mut T

Returns a raw pointer to the opaque data.

pub const fn raw_get(this: *const Self) -> *mut T

Gets the value behind this.

This function is useful to get access to the value without creating intermediate references.

impl<T> Opaque<T>

pub unsafe fn ffi_init(
init_func: unsafe extern "C" fn(_: *mut T)
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn ffi_init1<A1>(
init_func: unsafe extern "C" fn(_: *mut T, _: A1),
arg1: A1
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn ffi_init2<A1, A2>(
init_func: unsafe extern "C" fn(_: *mut T, _: A1, _: A2),
arg1: A1,
arg2: A2
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn ffi_init3<A1, A2, A3>(
init_func: unsafe extern "C" fn(_: *mut T, _: A1, _: A2, _: A3),
arg1: A1,
arg2: A2,
arg3: A3
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn ffi_init4<A1, A2, A3, A4>(
init_func: unsafe extern "C" fn(_: *mut T, _: A1, _: A2, _: A3, _: A4),
arg1: A1,
arg2: A2,
arg3: A3,
arg4: A4
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn manual_init(init_func: unsafe fn(_: *mut T)) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn manual_init1<A1>(
init_func: unsafe fn(_: *mut T, _: A1),
arg1: A1
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn manual_init2<A1, A2>(
init_func: unsafe fn(_: *mut T, _: A1, _: A2),
arg1: A1,
arg2: A2
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn manual_init3<A1, A2, A3>(
init_func: unsafe fn(_: *mut T, _: A1, _: A2, _: A3),
arg1: A1,
arg2: A2,
arg3: A3
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.

pub unsafe fn manual_init4<A1, A2, A3, A4>(
init_func: unsafe fn(_: *mut T, _: A1, _: A2, _: A3, _: A4),
arg1: A1,
arg2: A2,
arg3: A3,
arg4: A4
) -> impl PinInit<Self>

Create an initializer using the given initializer function.

Safety

The given function must under all circumstances initialize the memory location to a valid T. If it fails to do so it results in UB.

If any parameters are given, those need to be valid for the function. Valid means that calling the function with those parameters complies with the above requirement and every other requirement on the function itself.