Trait kernel::types::TrailingZeros

pub trait TrailingZeros {
    fn trailing_zeros(&self) -> u32;
}

Implemented by integer types that allow counting the number of trailing zeroes.

Required Methods

fn trailing_zeros(&self) -> u32

Returns the number of trailing zeroes in the binary representation of self.