Module kernel::user_ptr

User pointers.

C header: include/linux/uaccess.h

Structs

UserSlicePtr
A reference to an area in userspace memory, which can be either read-only or read-write.

UserSlicePtrReader
A reader for UserSlicePtr.

UserSlicePtrWriter
A writer for UserSlicePtr.

Traits

ReadableFromBytes
Specifies that a type is safely readable from byte slices.

WritableToBytes
Specifies that a type is safely writable to byte slices.