Struct kernel::utils::arcm::Arcm

pub struct Arcm<T: ?Sized> {
    arc: Arc<RawMutex<T, false>>,
}

Wrap the given type into an Arc and a Mutex. Arc allow multiple reference on the same data between threads Mutex allow any type to be Send and ensure safe access to the underlying data

Fields
arc: Arc<RawMutex<T, false>>
Implementations

impl<T> Arcm<T>

pub fn new(data: T) -> Self

Create a new Arcm by copying the data

Examples

let arcm: Arcm<u32> = Arcm::new(5);

pub fn execute<R>(
&self,
callback: impl FnMut(Arc<RawMutex<T, false>>) -> R
) -> R

Clone the current Arc and send it to the function pass in paramters

Examples

let arcm: Arcm<u32> = Arcm::new(5);
arcm.execute(|cloned| {
    *cloned.lock() = 10;
});
assert_eq!(*arcm, 10);