Struct kernel::utils::arcm::RawArcm

pub struct RawArcm<T: ?Sized, const INT: bool> {
    arc: Arc<RawMutex<T, INT>>,
}

Wrap the given type into an Arc and a Mutex. Arc allow multiple reference on the same data between threads Mutex allow any type to be Send and ensure safe access to the underlying data

Fields
arc: Arc<RawMutex<T, INT>>
Implementations

impl<T, const INT: bool> RawArcm<T, INT>

pub fn new(data: T) -> Self

Create a new RawArcm by copying the data

Examples

let arcm: RawArcm<u32, false> = RawArcm::new(5);

pub fn execute<R>(&self, callback: impl FnMut(Arc<RawMutex<T, INT>>) -> R) -> R

Clone the current Arc and send it to the function pass in paramters

Examples

let arcm: RawArcm<u32, false> = RawArcm::new(5);
arcm.execute(|cloned| {
    *cloned.lock() = 10;
});
assert_eq!(*arcm, 10);