Trait kernel::utils::flags::FlagOp

pub trait FlagOp<T> {
    fn toggle(&self, bit: u8);
    fn enable(&self, bit: u8);
    fn disable(&self, bit: u8);
    fn is_enable(&self, bit: u8) -> bool;
    fn is_disable(&self, bit: u8) -> bool;
}

To use this trait the type T has to allow interior mutability This is ok in the example since AtomicU8 is thread safe and allow interior mutability

Required Methods

fn toggle(&self, bit: u8)

fn enable(&self, bit: u8)

fn disable(&self, bit: u8)

fn is_enable(&self, bit: u8) -> bool

fn is_disable(&self, bit: u8) -> bool