Module kernel::x86_64

This module defines the kernel entry-point on x86_64.

Modules

acpi_handler

address_space

cpu

interrupts

logger

memory
This module contains the physical memory manager Pebble uses on x86_64.

task

Structs

Arch

Statics

GDT
The kernel GDT. This is not thread-safe, and so should only be altered by the bootstrap processor.

Functions

drop_to_userboot

kmain
This is the entry point for the kernel on x86_64. It is called from the UEFI bootloader and initialises the system, then passes control into the common part of the kernel.