Module kernel::x86_64::interrupts::exception

This module contains all the interrupt handlers used to handle CPU exceptions. Some of these exceptions are handled and recovered from, while some are fatal errors and lead to kernel panics.

Functions

breakpoint_handler

double_fault_handler

general_protection_fault_handler

invalid_opcode_handler

nmi_handler

page_fault_handler