Struct kernel::x86_64::memory::buddy_allocator::BuddyAllocator

pub struct BuddyAllocator {
    bins: Vec<BTreeSet<Frame>>,
}

Fields
bins: Vec<BTreeSet<Frame>>

The bins of free blocks, where bin i contains blocks of size 2^i. Uses BTreeSet to store the blocks in each bin, for efficient buddy location. The Frame stored for each block is the first frame in that block - the end frame can be calculated from the start frame and the order of the bin the block is in.

Methods

impl BuddyAllocator

pub fn new(max_order: usize) -> BuddyAllocator

Create a new BuddyAllocator, with a maximum block size of 2^max_order.

pub fn add_range(&mut self, range: Range<Frame>)

Add a range of Frames to this allocator, marking them free to allocate.

pub fn allocate_n(&mut self, n: usize) -> Option<Frame>

Allocate (at least) n contiguous frames from this allocator. Returns None if this allocator can't satisfy the allocation.

pub fn free_n(&mut self, start_frame: Frame, n: usize)

Free the given block (starting at start and of size n frames). n must be a power-of-2.

fn allocate_block(&mut self, order: usize) -> Option<Frame>

Tries to allocate a block of the given order. If no blocks of the correct size are available, tries to recursively split a larger block to form a block of the requested size.

fn buddy_of(x: Frame, order: usize) -> Frame

Finds the starting frame of the block that is the buddy of the block of order order, starting at x.

fn max_order(&self) -> usize

Get the order of the largest block this allocator can track.