Struct kernel::x86_64::per_cpu::PerCpu

pub struct PerCpu {
    _self_pointer: *const PerCpu,
    _pin: PhantomPinned,
    current_task_kernel_rsp: VirtualAddress,
    common: CommonPerCpu,
    tss: Tss,
    tss_selector: Option<SegmentSelector>,
}

Fields
_self_pointer: *const PerCpu

The first field of this structure must be a pointer to itself. This is because to access the per-cpu info, we read from gs:0x0, which is this pointer, and then dereference that to access the whole structure.

_pin: PhantomPinned

This structure must be pinned in memory for two reasons:

the self pointer makes this structure self-referential.

we put the address of this structure in the IA32_GS_BASE MSR. If this structure moves, that memory address becomes invalid and accessing the per-cpu data no longer has defined behaviour.

current_task_kernel_rsp: VirtualAddress

This holds the kernel rsp of the current task, and makes it efficient and easy to switch to the kernel stack upon kernel entry using syscall. It must remain at a fixed offset within this struct, because we refer to it manually with gs:0x8.

common: CommonPerCputss: Tsstss_selector: Option<SegmentSelector>
Methods

impl PerCpu

pub fn tss<'a>(self: Pin<&'a Self>) -> Pin<&'a Tss>

pub fn tss_mut<'a>(self: Pin<&'a mut Self>) -> Pin<&'a mut Tss>

pub fn current_task_kernel_rsp<'a>(self: Pin<&'a Self>) -> &'a VirtualAddress

pub fn current_task_kernel_rsp_mut<'a>(
self: Pin<&'a mut Self>
) -> Pin<&'a mut VirtualAddress>

pub fn common<'a>(self: Pin<&'a Self>) -> &'a CommonPerCpu

pub fn common_mut<'a>(self: Pin<&'a mut Self>) -> &'a mut CommonPerCpu