Crate kernel

Modules

arch

bit_vec
Collections implemented with bit vectors.

bitflags
A typesafe bitmask flag generator.

console

deferred

elf

io
Input and output.

ksyscall

lazy_static
A macro for declaring lazily evaluated statics.

libc_helpers

logging

multiboot

mutex

once
A "once initialization" primitive

phys_mem

prelude

process

ptr

singleton

tar

thread

unwind

virt_mem

Macros

bitflags!
The bitflags! macro generates a struct that holds a set of C-style bitmask flags. It is useful for creating typesafe wrappers for C APIs.

lazy_static!

lock!

Functions

kmain