Crate kernel

This module probably looks rather sparse! Check the root of one of the architecture modules for an entry point.

Modules

__rg_allocator_abi

arch

heap_allocator

object

syscall

util

x86_64
This module defines the kernel entry-point on x86_64.

Macros

assert_first_call
This macro should be called at the beginning of functions that create logic errors if they are called more than once. Most commonly this is used for initialization functions.

kernel_object_table

restore_regs
Restore the saved scratch registers.

save_regs
Macro to save the scratch registers. In System-V, rbx, rbp, r12, r13, r14, and r15 must be restored by the callee, so Rust automatically generates code to restore them, but for the rest we have to manually preserve them. Use restore_regs to restore the scratch registers before returning from the handler.

set_handler

set_handler_with_error_code

wrap_handler

wrap_handler_with_error_code

Statics

ALLOCATOR

Functions

kmain
This is the entry point for the kernel on x86_64. It is called from the UEFI bootloader and initialises the system, then passes control into the common part of the kernel.

panic