Crate kernel

The kernel crate.

This crate contains the kernel APIs that have been ported or wrapped for usage by Rust code in the kernel and is shared by all of them.

In other words, all the rest of the Rust code in the kernel (e.g. kernel modules written in Rust) depends on core, alloc and this crate.

If you need a kernel C API that is not ported or wrapped yet here, then do so first instead of bypassing this crate.

Re-exports

pub use macros;

pub use uapi;

Modules

error

Kernel errors.

init

API to safely and fallibly initialize pinned structs using in-place constructors.

ioctl

ioctl() number definitions

kunit

KUnit-based macros for Rust unit tests.

net

Networking.

prelude

The kernel prelude.

print

Printing facilities.

str

String representations.

sync

Synchronisation primitives.

task

Tasks (threads and processes).

types

Kernel types.

workqueue

Work queues.

Macros

b_str

Creates a new BStr from a string literal.

build_assert

Asserts that a boolean expression is true at compile time.

build_error

Fails the build if the code path calling build_error! can possibly be executed.

c_str

Creates a new CStr from a string literal.

current

Returns the currently running task.

dbg

std::dbg, but using pr_info instead of eprintln.

fmt

A convenience alias for core::format_args.

impl_has_work

Used to safely implement the HasWork<T, ID> trait.

init

Construct an in-place initializer for structs.

module_phy_driver

Declares a kernel module for PHYs drivers.

new_condvar

Creates a CondVar initialiser with the given name and a newly-created lock class.

new_mutex

Creates a Mutex initialiser with the given name and a newly-created lock class.

new_spinlock

Creates a SpinLock initialiser with the given name and a newly-created lock class.

new_work

Creates a Work initialiser with the given name and a newly-created lock class.

pin_init

Construct an in-place, pinned initializer for structs.

pr_alert

Prints an alert-level message (level 1).

pr_cont

Continues a previous log message in the same line.

pr_crit

Prints a critical-level message (level 2).

pr_debug

Prints a debug-level message (level 7).

pr_emerg

Prints an emergency-level message (level 0).

pr_err

Prints an error-level message (level 3).

pr_info

Prints an info-level message (level 6).

pr_notice

Prints a notice-level message (level 5).

pr_warn

Prints a warning-level message (level 4).

stack_pin_init

Initialize and pin a type directly on the stack.

stack_try_pin_init

Initialize and pin a type directly on the stack.

static_assert

Static assert (i.e. compile-time assert).

try_init

Construct an in-place fallible initializer for structs.

try_pin_init

Construct an in-place, fallible pinned initializer for structs.

Structs

ThisModule

Equivalent to THIS_MODULE in the C API.

Traits

Module

The top level entrypoint to implementing a kernel module.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.