Crate kernel

Reexports

extern crate core as core;

use core::prelude::v1::*;

pub use callback::AppId;

pub use callback::Callback;

pub use container::Container;

pub use driver::Driver;

pub use mem::AppSlice;

pub use mem::AppPtr;

pub use mem::Private;

pub use mem::Shared;

pub use process::Process;

pub use process::State;

pub use returncode::ReturnCode;

Modules

callback
Data structure for storing a callback to userspace or kernelspace.

common
Common operations in the Tock OS.

container
Data structure to store a list of userspace applications.

debug
Provides a debug! macro for in-kernel debugging.

driver
System call interface for userspace applications.

hil
Public traits for interfaces between Tock components.

ipc
Inter-process communication mechanism for Tock.

mem
Data structure for passing application memory to the kernel.

memop
Implementation of the MEMOP family of syscalls.

mpu
Interface for configuring the Memory Protection Unit.

platform

process
Support for creating and running userspace applications.

returncode
Standard return type for invoking operations, returning success or an error code.

sched
Tock core scheduler.

support
A bare-metal library supplying functions rustc may lower code to.

syscall
Tock syscall number definitions.

systick
Interface and default implementation for the system tick timer.

Macros

debug
In-kernel printf() debugging.

static_init
Allocates a global array of static size to initialize data structures.

Traits

Chip
Interface for individual MCUs.

Platform
Interface for individual boards.

SysTick
Interface for the system tick timer.

Functions

main
Main loop.