Crate kernel

Re-exports

pub use config::config;

pub use userspace::pid::KernelPid;

Modules

config

drivers

framebuffer

gdt

heap

init

initramfs

interrupts

iobus

This entire file is largely based on the examples in https://wiki.osdev.org/PIC#Programming_the_PIC_chips

ipc

memory

panic

pic

This entire file is largely based on the examples in https://wiki.osdev.org/PIC#Programming_the_PIC_chips

scheduler

userspace

Macros

debug

emergency

error

info

kcolorchange

kprint

kprintln

———————————— Macros ––––––––––––––––––––

kscopedcolorchange

success

warn

Structs

Handover

MessagePointer2

Pid

Process

Constants

HEAP_ADDRESS

HEAP_LENGTH

Functions

kmain

Unions

MessageContent