Trait kernel::devices::clock::ClockDevice

trait ClockDevice: Send + Sync {
    // Required methods
    fn name(&self) -> &'static str;
    fn get_time(&self) -> ClockTime;
    fn granularity(&self) -> u64;
    fn require_calibration(&self) -> bool;

    // Provided method
    fn rating(&self) -> u64 { ... }
}

Required Methods

fn name(&self) -> &'static str

Returns the name of the device

fn get_time(&self) -> ClockTime

Returns the current time of the device with no relation to anything The system will use consequtive calls to determine the time

fn granularity(&self) -> u64

Returns the granularity of the device in nanoseconds, i.e. the smallest time unit it can measure Must be at least 1

fn require_calibration(&self) -> bool

Returns true if the device needs to be calibration i.e. it doesn’t count time correctly

Provided Methods

fn rating(&self) -> u64

Returns the rating of the device, i.e. how good it is The higher the better