Module kernel::file

Files and file descriptors.

C headers: include/linux/fs.h and include/linux/file.h

Modules

flags

Flags associated with a File.

Structs

File

Wraps the kernel’s struct file.

FileDescriptorReservation

A file descriptor reservation.

IoctlCommand

Represents an ioctl command.

PollTable

Wraps the kernel’s struct poll_table_struct.

Enums

SeekFrom

Equivalent to std::io::SeekFrom.

Traits

IoctlHandler

Allows the handling of ioctls defined with the _IO, _IOR, _IOW, and _IOWR macros.

OpenAdapter

Trait for extracting file open arguments from kernel data structures.

Operations

Corresponds to the kernel’s struct file_operations.