kernel

Module fs

Modules

fat 🔒

mapping

mbr 🔒

path

A library for parsing, manipulating Paths that follow unix style conventions.

Structs

AccessHelper

BaseNode

Directory

A handle to a directory, it has the inode which controls the properties of the node in the filesystem

DirectoryNode

EmptyFileSystem

File

A handle to a file, it has the inode which controls the properties of the node in the filesystem

FileAccess

FileAttributes

FileNode

Enums

DirTraverse

FileSystemError

FilesystemNode

A node in the filesystem, can be a file or a directory

Node

A node of the filesystem, it can be anything, a file, a device or a directory

Constants

ANOTHER_FILESYSTEM_MAPPING_INODE_MAGIC 🔒

This is not used at all, just an indicator in Directory::fetch_entries

NO_PARENT_DIR_SECTOR 🔒

Statics

EMPTY_FILESYSTEM 🔒

Traits

FileSystem

A filesystem trait, this is the main interface to interact with the filesystem it is used to open files, directories, read and write files, etc.

Functions

create_disk_mapping

Loads the hard disk specified in the argument it will load the first partition (MBR) if any, otherwise it will treat the whole disk as one partition

empty_filesystem

open_inode 🔒

Open the inode of a path, this include directories and files.

unmount_all