Module kernel::fs

Re-exports

pub use file::*;

Modules

file 🔒

Statics

SYSFILES 🔒

Functions

close

Close a file given its file descriptor. This does not delete the file from the system

create

Take information on a file and add it to SYSFILES if it does not exist ErrNo is return if file already found.

create_from_raw

Create a file given its name and a predefined buffer. The buffer should implement FileOperation trait. WARNINGS: This does use String and Box to create FileInfo, no file can be created before Heap creation

delete

Delete file from SYSFILE given its name Should be updated later to check permission on the file

open

Look for a file given its name in SYSFILES and open it. Open files list is common between processses, this will change in later version

read

This function mimic the linux read syscall. Look for a file in file lists and call it’s FileOperation implementation. Mutex on PROC_FILES is acquire during all the read processus which imply you can’t r/w another file at the same time.

socket_pair

Create and open a pair of socket given it’s domain, type and protocol. Fd are written to sockets array. Prototype is made to match linux syscall

write

This function mimic the linux write syscall. Look for a file in file lists and call it’s FileOperation implementation. Mutex on PROC_FILES is acquire during all the read processus which imply you can’t r/w another file at the same time.