Struct kernel::fs::file::socket::Socket

pub struct Socket {
    domain: SocketDomain,
    stype: SocketType,
    protocol: SocketProtocol,
    buffer: Option<[RawArcm<RawFileMemory, false>; 2]>,
    endpoint: usize,
}

Socket structure representation. Socket alone can’t do much. These need to be created by pair, Each socket will be tide to both endpoint but we’ll access only one by writing or reading. UNIX domain will create 2 buffers for both endpoint of the socket. INET domain is not implemented but could create an endpoint of the socket bound to the network interface TODO? Maybe embed the woffset inside the buffer to precisely know how much byte as been written

Fields
domain: SocketDomainstype: SocketTypeprotocol: SocketProtocolbuffer: Option<[RawArcm<RawFileMemory, false>; 2]>endpoint: usize
Implementations

impl Socket

pub const fn new( domain: SocketDomain, stype: SocketType, protocol: SocketProtocol ) -> Self

Create a socket given it’s domain, type and protocol

impl Socket

FileOperations for DGRAM sockets

fn dgram_read(&self, dst: &mut [u8], length: usize) -> Result<usize, ErrNo>

fn dgram_write(&mut self, src: &[u8], length: usize) -> Result<usize, ErrNo>

impl Socket

FileOperations for STREAM sockets

fn stream_read(&self, _dst: &mut [u8], _length: usize) -> Result<usize, ErrNo>

fn stream_write(&mut self, _src: &[u8], _length: usize) -> Result<usize, ErrNo>

impl Socket

FileOperations for RAW sockets

fn raw_read(&self, _dst: &mut [u8], _length: usize) -> Result<usize, ErrNo>

fn raw_write(&mut self, _src: &[u8], _length: usize) -> Result<usize, ErrNo>