Module kernel::fs::flags

File system flags.

Constants

ALLOW_IDMAP

The file system has been updated to handle vfs idmappings.

BINARY_MOUNTDATA

The options provided when mounting are in binary form.

DISALLOW_NOTIFY_PERM

Disables fanotify permission events.

HAS_SUBTYPE

The file system has a subtype. It is extracted from the name and passed in as a parameter.

RENAME_DOES_D_MOVE

The file systen will handle d_move during rename internally.

REQUIRES_DEV

The file system requires a device.

USERNS_MOUNT

The file system can be mounted by userns root.

This documentation is an old archive. Please see https://rust.docs.kernel.org instead.