Trait kernel::fs::Type

pub trait Type {
    type Context: Context<Self> + ?Sized;

    const NAME: &'static CStr;
    const MAGIC: u32;
    const FLAGS: i32;
}

A file system type.

Required Associated Types

type Context: Context<Self> + ?Sized

The context used to build fs configuration before it is mounted or reconfigured.

Required Associated Constants

const NAME: &'static CStr

The name of the file system type.

const MAGIC: u32

The magic number associated with the file system.

This is normally one of the values in include/uapi/linux/magic.h.

const FLAGS: i32

The flags of this file system type.

It is a combination of the flags in the flags module.