Trait kernel::gpio::Chip

pub trait Chip {
    type Data: PointerWrapper + Sync + Send;

    const TO_USE: ToUse;
    fn get_direction(
        _data: <Self::Data as PointerWrapper>::Borrowed, 
        _offset: u32
    ) -> Result<LineDirection> { ... }

    fn direction_input(
        _data: <Self::Data as PointerWrapper>::Borrowed, 
        _offset: u32
    ) -> Result { ... }

    fn direction_output(
        _data: <Self::Data as PointerWrapper>::Borrowed, 
        _offset: u32, 
        _value: bool
    ) -> Result { ... }

    fn get(
        _data: <Self::Data as PointerWrapper>::Borrowed, 
        _offset: u32
    ) -> Result<bool> { ... }

    fn set(
        _data: <Self::Data as PointerWrapper>::Borrowed, 
        _offset: u32, 
        _value: bool
    ) { ... }
}

A gpio chip.

Associated Types

type Data: PointerWrapper + Sync + Send

Context data associated with the gpio chip.

It determines the type of the context data passed to each of the methods of the trait.

Associated Constants

const TO_USE: ToUse

The methods to use to populate [struct gpio_chip]. This is typically populated with declare_gpio_chip_operations.

Provided methods

fn get_direction(
_data: <Self::Data as PointerWrapper>::Borrowed,
_offset: u32
) -> Result<LineDirection>

Returns the direction of the given gpio line.

fn direction_input(
_data: <Self::Data as PointerWrapper>::Borrowed,
_offset: u32
) -> Result

Configures the direction as input of the given gpio line.

fn direction_output(
_data: <Self::Data as PointerWrapper>::Borrowed,
_offset: u32,
_value: bool
) -> Result

Configures the direction as output of the given gpio line.

The value that will be initially output is also specified.

fn get(
_data: <Self::Data as PointerWrapper>::Borrowed,
_offset: u32
) -> Result<bool>

Returns the current value of the given gpio line.

fn set(
_data: <Self::Data as PointerWrapper>::Borrowed,
_offset: u32,
_value: bool
)

Sets the value of the given gpio line.