Module kernel::heap_allocator

Structs

Allocation 🔒

Hole

HoleAllocator

HoleInfo

HoleList

LockedHoleAllocator

Functions

allocate_first_fit 🔒

free 🔒

Walk the list, starting at hole and free the allocation given by (addr,size)

handle_alloc_error 🔒

split_hole 🔒

Split the given hole into (front_padding,hole,back_padding) if it’s big enough to hold the given layout with the required alignment. - Front padding occurs when the required alignment is higher than that of the hole. - Back padding occurs when the layout’s size is smaller than the hole.