Trait kernel::hwrng::Operations

pub trait Operations {
    type Data: PointerWrapper + Send + Sync = ();

    const TO_USE: ToUse;

    fn read(
        data: <Self::Data as PointerWrapper>::Borrowed, 
        buffer: &mut [u8], 
        wait: bool
    ) -> Result<u32>;

    fn init(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result { ... }

    fn cleanup(_data: Self::Data) { ... }
}

This trait is implemented in order to provide callbacks to struct hwrng.

Associated Types

type Data: PointerWrapper + Send + Sync = ()

The pointer type that will be used to hold user-defined data type.

Associated Constants

const TO_USE: ToUse

The methods to use to populate [struct hwrng].

Required methods

fn read(
data: <Self::Data as PointerWrapper>::Borrowed,
buffer: &mut [u8],
wait: bool
) -> Result<u32>

Read data into the provided buffer. Drivers can fill up to max bytes of data into the buffer. The buffer is aligned for any type and its size is a multiple of 4 and >= 32 bytes.

Provided methods

fn init(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result

Initialization callback, can be left undefined.

fn cleanup(_data: Self::Data)

Cleanup callback, can be left undefined.