Trait kernel::init::Zeroable

pub unsafe trait Zeroable { }

Marker trait for types that can be initialized by writing just zeroes.

Safety

The bit pattern consisting of only zeroes is a valid bit pattern for this type. In other words, this is not UB:

let val: Self = unsafe { core::mem::zeroed() };