Module kernel::irq

Interrupts and interrupt chips.

See https://www.kernel.org/doc/Documentation/core-api/genericirq.rst.

C headers: include/linux/irq.h and include/linux/interrupt.h.

Structs

ChainedGuard

A guard to call chained_irq_exit after chained_irq_enter was called.

Descriptor

Wraps the kernel’s struct irq_desc.

Domain

Wraps the kernel’s struct irq_domain.

IrqData

Wraps the kernel’s struct irq_data.

LockedIrqData

Wraps the kernel’s struct irq_data when it is locked.

ToUse

Represents which fields of [struct irq_chip] should be populated with pointers.

Type

Contains constants that describes how an interrupt can be triggered.

Enums

ExtraResult

Extra information returned by some of the Chip methods on success.

Constants

USE_NONE

A constant version where all values are to set to false, that is, all supported fields will be set to null pointers.

Traits

Chip

An irq chip.

FlowHandler

A high-level irq flow handler.

Functions

set_wake

Enables or disables power-management wake-on for the given irq number.

Type Definitions

HwNumber

The type of irq hardware numbers.