Struct kernel::memory_management::virtual_memory_mapper::VirtualMemoryMapper

pub struct VirtualMemoryMapper {
    page_map_l4: PageDirectoryTablePtr,
    is_user: bool,
}

Fields
page_map_l4: PageDirectoryTablePtris_user: bool
Implementations

impl VirtualMemoryMapper

const fn boot_vm() -> Self

Return the VM for the CPU at boot time (only applied to the first CPU and this is setup in boot.S)

fn new() -> Self

pub fn clone_kernel_mem(&self) -> Self

pub unsafe fn add_process_specific_mappings(&mut self)

Safety

After this call, the VM must never be switched to unless its from the scheduler or we are sure that the previous kernel regions are not used

fn load_vm(base: &PageDirectoryTablePtr)

fn get_current_vm() -> Self

pub unsafe fn switch_to_this(&self)

Safety

This must be used with caution, it must never be switched while we are using memory from the same regions, i.e. kernel stack while we are in an interrupt

fn new_kernel_vm() -> Self

pub fn map(&mut self, entry: &VirtualMemoryMapEntry)

pub fn unmap(&mut self, entry: &VirtualMemoryMapEntry, is_allocated: bool)

Removes mapping of a virtual entry, it will free it from physical memory if it was allocated

pub fn is_address_mapped(&self, addr: u64) -> bool

fn do_for_ranges_enteries<R1, R2, F>( &mut self, l4_ranges: R1, l3_ranges: R2, f: F )
where R1: RangeBounds<usize>, R2: RangeBounds<usize>, F: FnMut(&mut u64),

fn do_for_every_user_entry(&mut self, f: impl FnMut(&mut u64))

fn do_for_kernel_process_entry(&mut self, f: impl FnMut(&mut u64))

pub fn unmap_process_memory(&mut self)