Struct kernel::object::memory_object::MemoryObject

pub struct MemoryObject {
    pub id: KernelObjectId,
    pub owner: KernelObjectId,
    pub virtual_address: Option<VirtualAddress>,
    pub physical_address: PhysicalAddress,
    pub size: usize,
    pub flags: Flags,
}

Fields
id: KernelObjectIdowner: KernelObjectIdvirtual_address: Option<VirtualAddress>

The virtual address to map this MemoryObject at. If this is None, the mapping task can choose to map it at any virtual address it chooses.

physical_address: PhysicalAddresssize: usize

Size of this MemoryObject in bytes.

flags: Flags
Implementations

impl MemoryObject

pub fn new(
owner: KernelObjectId,
virtual_address: Option<VirtualAddress>,
physical_address: PhysicalAddress,
size: usize,
flags: Flags
) -> Arc<MemoryObject>

pub fn from_boot_info(
owner: KernelObjectId,
segment: &Segment
) -> Arc<MemoryObject>