Trait kernel::platform::Driver

pub trait Driver {
    type Data: PointerWrapper + Send + Sync + DeviceRemoval = ();
    type IdInfo: 'static = ();

    const OF_DEVICE_ID_TABLE: Option<IdTable<'static, DeviceId, Self::IdInfo>> = None;

    fn probe(
        dev: &mut Device, 
        id_info: Option<&Self::IdInfo>
    ) -> Result<Self::Data>;

    fn remove(_data: &Self::Data) -> Result { ... }
}

A platform driver.

Provided Associated Types

type Data: PointerWrapper + Send + Sync + DeviceRemoval = ()

Data stored on device by driver.

Corresponds to the data set or retrieved via the kernel’s platform_{set,get}_drvdata() functions.

Require that Data implements PointerWrapper. We guarantee to never move the underlying wrapped data structure. This allows

type IdInfo: 'static = ()

The type holding information about each device id supported by the driver.

Provided Associated Constants

const OF_DEVICE_ID_TABLE: Option<IdTable<'static, DeviceId, Self::IdInfo>> = None

The table of device ids supported by the driver.

Required Methods

fn probe(dev: &mut Device, id_info: Option<&Self::IdInfo>) -> Result<Self::Data>

Platform driver probe.

Called when a new platform device is added or discovered. Implementers should attempt to initialize the device here.

Provided Methods

fn remove(_data: &Self::Data) -> Result

Platform driver remove.

Called when a platform device is removed. Implementers should prepare the device for complete removal here.