Trait kernel::power::Operations

pub trait Operations {
    type Data: PointerWrapper + Sync + Send;
    fn suspend(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result { ... }

    fn resume(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result { ... }

    fn freeze(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result { ... }

    fn restore(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result { ... }
}

Corresponds to the kernel’s struct dev_pm_ops.

It is meant to be implemented by drivers that support power-management operations.

Associated Types

type Data: PointerWrapper + Sync + Send

The type of the context data stored by the driver on each device.

Provided methods

fn suspend(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result

Called before the system goes into a sleep state.

fn resume(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result

Called after the system comes back from a sleep state.

fn freeze(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result

Called before creating a hibernation image.

fn restore(_data: <Self::Data as PointerWrapper>::Borrowed) -> Result

Called after the system is restored from a hibernation image.