Module kernel::prelude

The kernel prelude.

These are the most common items used by Rust code in the kernel, intended to be imported by all Rust code, for convenience.

Examples

use kernel::prelude::*;

Re-exports

pub use super::build_assert;

pub use super::dev_alert;

pub use super::dev_crit;

pub use super::dev_dbg;

pub use super::dev_emerg;

pub use super::dev_err;

pub use super::dev_info;

pub use super::dev_notice;

pub use super::dev_warn;

pub use super::pr_alert;

pub use super::pr_crit;

pub use super::pr_debug;

pub use super::pr_emerg;

pub use super::pr_err;

pub use super::pr_info;

pub use super::pr_notice;

pub use super::pr_warn;

pub use super::module_misc_device;

pub use super::static_assert;

pub use super::KernelModule;

pub use super::str::CStr;

pub use super::ThisModule;

Macros

dbg

std::dbg, but using pr_info instead of eprintln.

module

Declares a kernel module.

Structs

Box

A pointer type for heap allocation.

Error

Generic integer kernel error.

Pin

A pinned pointer.

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T> and pronounced ‘vector’.

Type Definitions

Result

A Result with an Error error type.