Module kernel::prelude

The kernel prelude.

These are the most common items used by Rust code in the kernel, intended to be imported by all Rust code, for convenience.

Examples

use kernel::prelude::*;

Re-exports

pub use core::pin::Pin;

pub use alloc::boxed::Box;

pub use alloc::vec::Vec;

pub use macros::module;

pub use macros::vtable;

pub use super::build_assert;

pub use super::dbg;

pub use super::pr_alert;

pub use super::pr_crit;

pub use super::pr_debug;

pub use super::pr_emerg;

pub use super::pr_err;

pub use super::pr_info;

pub use super::pr_notice;

pub use super::pr_warn;

pub use super::static_assert;

pub use super::error::code::*;

pub use super::error::Error;

pub use super::error::Result;

pub use super::str::CStr;

pub use super::ThisModule;