Module kernel::prelude

The kernel prelude.

These are the most common items used by Rust code in the kernel, intended to be imported by all Rust code, for convenience.

Examples

use kernel::prelude::*;

Re-exports

pub use core::pin::Pin;

pub use alloc::boxed::Box;

pub use alloc::vec::Vec;

pub use macros::module;

pub use macros::pin_data;

pub use macros::pinned_drop;

pub use macros::vtable;

pub use super::dbg;

pub use super::init;

pub use super::init;

pub use super::error::Error;

pub use super::error::Result;

pub use super::str::CStr;

pub use super::ThisModule;

pub use super::init::InPlaceInit;

pub use super::init::Init;

pub use super::init::PinInit;

pub use super::error::code::*;

Macros

build_assert

Asserts that a boolean expression is true at compile time.

current

Returns the currently running task.

init

Construct an in-place initializer for structs.

pin_init

Construct an in-place, pinned initializer for structs.

pr_alert

Prints an alert-level message (level 1).

pr_crit

Prints a critical-level message (level 2).

pr_debug

Prints a debug-level message (level 7).

pr_emerg

Prints an emergency-level message (level 0).

pr_err

Prints an error-level message (level 3).

pr_info

Prints an info-level message (level 6).

pr_notice

Prints a notice-level message (level 5).

pr_warn

Prints a warning-level message (level 4).

static_assert

Static assert (i.e. compile-time assert).

try_init

Construct an in-place fallible initializer for structs.

try_pin_init

Construct an in-place, fallible pinned initializer for structs.