Struct kernel::process::Process

pub struct Process<'a> {
    memory: &'static mut [u8],
    kernel_memory_break: *const u8,
    app_heap_break: *const u8,
    app_heap_start: *const u8,
    stack_data_boundary: *const u8,
    cur_stack: *const u8,
    min_stack_pointer: *const u8,
    syscall_count: Cell<usize>,
    last_syscall: Cell<Option<Syscall>>,
    text: &'static [u8],
    stored_regs: StoredRegs,
    yield_pc: usize,
    psr: usize,
    state: State,
    fault_response: FaultResponse,
    mpu_regions: [Cell<(*const u8, PowerOfTwo)>; 5],
    tasks: RingBuffer<'a, Task>,
    pub package_name: &'static str,
}

Fields
memory: &'static mut [u8]

Application memory layout:

|======== <- memory[memory.len()]

╔═ | Grant | ↓ D | ---- <- kernel_memory_break Y | N | ---- <- app_heap_break A | M | ↑ | Heap ╠═ | ---- <- app_heap_start | Data F | ---- <- stack_data_boundary I | Stack X | ↓ E | D | ---- <- cur_stack | ╚═ |======== <- memory[0] The process's memory.

kernel_memory_break: *const u8 app_heap_break: *const u8 app_heap_start: *const u8 stack_data_boundary: *const u8 cur_stack: *const u8 min_stack_pointer: *const u8

How low have we ever seen the stack pointer

syscall_count: Cell<usize>

How many syscalls have occurred since the process started

last_syscall: Cell<Option<Syscall>>

What was the most recent syscall

text: &'static [u8]

Process text segment

stored_regs: StoredRegs yield_pc: usize psr: usize state: State fault_response: FaultResponse

How to deal with Faults occuring in the process

mpu_regions: [Cell<(*const u8, PowerOfTwo)>; 5]

MPU regions are saved as a pointer-size pair.

size is encoded as X where SIZE = 2X + 1 and X >= 4.

A null pointer represents an empty region.

Invariants

The pointer must be aligned to the size. E.g. if the size is 32 bytes, the pointer must be 32-byte aligned.

tasks: RingBuffer<'a, Task> package_name: &'static str
Methods

impl<'a> Process<'a>

fn schedule_ipc(&mut self, from: AppId, cb_type: IPCType)

fn current_state(&self) -> State

fn yield_state(&mut self)

unsafe fn fault_state(&mut self)

fn dequeue_task(&mut self) -> Option<Task>

fn mem_start(&self) -> *const u8

fn mem_end(&self) -> *const u8

fn flash_start(&self) -> *const u8

fn flash_end(&self) -> *const u8

fn kernel_memory_break(&self) -> *const u8

fn setup_mpu<MPU: MPU>(&self, mpu: &MPU)

fn add_mpu_region(&self, base: *const u8, size: u32) -> bool

unsafe fn create(
app_flash_address: *const u8,
remaining_app_memory: *mut u8,
remaining_app_memory_size: usize,
fault_response: FaultResponse
) -> (Option<Process<'a>>, usize, usize)

fn sbrk(&mut self, increment: isize) -> Result<*const u8, Error>

fn brk(&mut self, new_break: *const u8) -> Result<*const u8, Error>

fn in_exposed_bounds(&self, buf_start_addr: *const u8, size: usize) -> bool

unsafe fn alloc(&mut self, size: usize) -> Option<&mut [u8]>

unsafe fn free<T>(&mut self, _: *mut T)

unsafe fn container_ptr<T>(&self, container_num: usize) -> *mut *mut T

unsafe fn container_for<T>(&mut self, container_num: usize) -> *mut T

unsafe fn container_for_or_alloc<T: Default>(
&mut self,
container_num: usize
) -> Option<*mut T>

fn pop_syscall_stack(&mut self)

unsafe fn push_function_call(&mut self, callback: FunctionCall)

Context switch to the process.

unsafe fn app_fault(&self) -> bool

unsafe fn syscall_fired(&self) -> bool

unsafe fn switch_to(&mut self)

Context switch to the process.

fn svc_number(&self) -> Option<Syscall>

fn incr_syscall_count(&self)

fn sp(&self) -> usize

fn lr(&self) -> usize

fn pc(&self) -> usize

fn r0(&self) -> usize

fn set_return_code(&mut self, return_code: ReturnCode)

fn set_r0(&mut self, val: isize)

fn r1(&self) -> usize

fn r2(&self) -> usize

fn r3(&self) -> usize

fn r12(&self) -> usize

fn xpsr(&self) -> usize

unsafe fn fault_str<W: Write>(&mut self, writer: &mut W)

unsafe fn statistics_str<W: Write>(&mut self, writer: &mut W)