Module kernel::sync

Synchronisation primitives.

This module contains the kernel APIs related to synchronisation that have been ported or wrapped for usage by Rust code in the kernel.

Re-exports

pub use lock::mutex::Mutex;

pub use lock::spinlock::SpinLock;

Modules

lock

Generic kernel lock and guard.

Structs

Arc

A reference-counted pointer to an instance of T.

ArcBorrow

A borrowed reference to an Arc instance.

CondVar

A conditional variable.

LockClassKey

Represents a lockdep class. It’s a wrapper around C’s lock_class_key.

LockedBy

Allows access to some data to be serialised by a lock that does not wrap it.

UniqueArc

A refcounted object that is known to have a refcount of 1.