Struct kernel::sync::Ref

pub struct Ref<T: ?Sized> { /* fields omitted */ }

A reference-counted pointer to an instance of T.

The reference count is incremented when new instances of Ref are created, and decremented when they are dropped. When the count reaches zero, the underlying T is also dropped.

Invariants

The reference count on an instance of Ref is always non-zero. The object pointed to by Ref is always pinned.

Implementations

impl<T> Ref<T>

pub fn try_new(contents: T) -> Result<Self>

Constructs a new reference counted instance of T.

pub fn try_new_and_init<U: FnOnce(Pin<&mut T>)>(
contents: T,
init: U
) -> Result<Self>

Constructs a new reference counted instance of T and calls the initialisation function.

This is useful because it provides a mutable reference to T at its final location.

pub fn into_usize(obj: Self) -> usize

Deconstructs a Ref object into a usize.

It can be reconstructed once via Ref::from_usize.

pub unsafe fn borrow_usize(encoded: usize) -> RefBorrow<T>

Borrows a Ref instance previously deconstructed via Ref::into_usize.

Safety

encoded must have been returned by a previous call to Ref::into_usize. Additionally, Ref::from_usize can only be called after all instances of RefBorrow have been dropped.

pub unsafe fn from_usize(encoded: usize) -> Self

Recreates a Ref instance previously deconstructed via Ref::into_usize.

Safety

encoded must have been returned by a previous call to Ref::into_usize. Additionally, it can only be called once for each previous call to ``Ref::into_usize`.

impl<T: ?Sized> Ref<T>

pub fn ptr_eq(a: &Self, b: &Self) -> bool

Determines if two reference-counted pointers point to the same underlying instance of T.

pub fn pinned(obj: Self) -> Pin<Self>

Returns a pinned version of a given Ref instance.