Struct kernel::sync::SpinLock

pub struct SpinLock<T: ?Sized> { /* fields omitted */ }

Exposes the kernel’s spinlock_t. When multiple CPUs attempt to lock the same spinlock, only one at a time is allowed to progress, the others will block (spinning) until the spinlock is unlocked, at which point another CPU will be allowed to make progress.

A SpinLock must first be initialised with a call to SpinLock::init before it can be used. The spinlock_init macro is provided to automatically assign a new lock class to a spinlock instance.

SpinLock does not manage the interrupt state, so it can be used in only two cases: (a) when the caller knows that interrupts are disabled, or (b) when callers never use it in interrupt handlers (in which case it is ok for interrupts to be enabled).

Implementations

impl<T> SpinLock<T>

pub unsafe fn new(t: T) -> Self

Constructs a new spinlock.

Safety

The caller must call SpinLock::init before using the spinlock.

impl<T: ?Sized> SpinLock<T>

pub fn lock(&self) -> Guard<'_, Self>

Locks the spinlock and gives the caller access to the data protected by it. Only one thread at a time is allowed to access the protected data.