Trait kernel::sync::NeedsLockClass

pub trait NeedsLockClass {
    unsafe fn init(
        self: Pin<&mut Self>, 
        name: &'static CStr, 
        key1: *mut lock_class_key, 
        key2: *mut lock_class_key
    );
}

A trait for types that need a lock class during initialisation.

Implementers of this trait benefit from the [init_with_lockdep] macro that generates a new class for each initialisation call site.

Required methods

unsafe fn init(
self: Pin<&mut Self>,
name: &'static CStr,
key1: *mut lock_class_key,
key2: *mut lock_class_key
)

Initialises the type instance so that it can be safely used.

Callers are encouraged to use the [init_with_lockdep] macro as it automatically creates a new lock class on each usage.

Safety

key1 and key2 must point to valid memory locations and remain valid until self is dropped.