kernel
Module syscall

Functions

create_address_space

handle_syscall

This is the architecture-independent syscall handler. It should be called by the handler that receives the syscall (each architecture is free to do this however it wishes). The only parameter that is guaranteed to be valid is number; the meaning of the rest may be undefined depending on how many parameters the specific system call takes.

poll_interest

resize_memory_object

spawn_task

wait_for_event