Struct kernel::system::System

pub struct System { /* private fields */ }

A Kernel of MEG-OS codename Maystorm

Implementations

impl System

pub unsafe fn init(info: &BootInfo, f: fn()) -> !

Initialize the system

pub const fn name() -> &'static str

Returns the name of the current system.

pub const fn codename() -> &'static str

Returns the codename of the current system.

pub const fn short_name() -> &'static str

Returns abbreviated name of the current system.

pub const fn version<'a>() -> &'a Version<'a>

Returns the version of the current system.

pub fn boot_flags() -> BootFlags

pub fn system_time() -> SystemTime

Returns the current system time.

pub const fn is_multi_processor_capable_kernel() -> bool

Returns whether the kernel is multiprocessor-capable.

pub unsafe fn activate_cpu(new_cpu: Box<Cpu>)

Add SMP-initialized CPU cores to the list of activated cores.

Safety

THREAD UNSAFE. Do not call this function except when initializing the SMP.

pub fn cpus<'a>() -> impl ExactSizeIterator<Item = &'a Box<Cpu>>

pub fn cpu<'a>(index: ProcessorIndex) -> &'a Cpu

Returns a reference to the processor at the specified index.

Panics

Panics if specified index is larger than the number of processors.

pub fn cpu_ref<'a>(index: ProcessorIndex) -> Option<&'a Box<Cpu>>

pub fn smbios<'a>() -> Option<&'a SmBios>

pub fn current_device<'a>() -> &'a DeviceInfo

Returns the current device information.

pub fn acpi<'a>() -> Option<&'a Xsdt>

pub fn safe_screen<'a>() -> Option<Arc<BitmapScreen<'static>>>

pub fn main_screen( ) -> Option<Arc<dyn Screen<BitmapRef32<'static>, ColorType = TrueColor>>>

Get main screen

pub fn set_stdout(stdout: Box<dyn Tty>)

pub fn stdout<'a>() -> &'a mut dyn Tty

pub fn log<'a>() -> &'a mut dyn Tty

pub fn assert_call_once(mutex: &'static AtomicBool)