Trait kernel::Platform

pub trait Platform: Sized + 'static {
    type PageTableSize: FrameSize;
    type PageTable: PageTable<Self::PageTableSize> + Send;
    type PerCpu: PerCpu<Self>;
    fn kernel_page_table(&mut self) -> &mut Self::PageTable;

    fn per_cpu<'a>() -> Pin<&'a mut Self::PerCpu>;

    unsafe fn initialize_task_kernel_stack(
        kernel_stack_top: &mut VirtualAddress, 
        task_entry_point: VirtualAddress, 
        user_stack_top: &mut VirtualAddress
    );

    unsafe fn context_switch(
        current_kernel_stack: *mut VirtualAddress, 
        new_kernel_stack: VirtualAddress
    );

    unsafe fn drop_into_userspace(kernel_stack_pointer: VirtualAddress) -> !;
}

Associated Types

type PageTableSize: FrameSize

type PageTable: PageTable<Self::PageTableSize> + Send

type PerCpu: PerCpu<Self>

Loading content...
Required methods

fn kernel_page_table(&mut self) -> &mut Self::PageTable

fn per_cpu<'a>() -> Pin<&'a mut Self::PerCpu>

Get the per-CPU info for the current CPU. To make this safe, the per-CPU info must be installed before the Platform implementation is created.

unsafe fn initialize_task_kernel_stack(
kernel_stack_top: &mut VirtualAddress,
task_entry_point: VirtualAddress,
user_stack_top: &mut VirtualAddress
)

Often, the kernel stack of a task must be initialized to allow it to enter usermode for the first time. What is required for this is architecture-dependent, and so this is offloaded to the TaskHelper.

entry_point is the address that should be jumped to in usermode when the task is run for the first time. user_stack_top is the virtual address that should be put into the stack pointer when the task is entered.

kernel_stack_top is the kernel stack that the new stack frames will be installed in, and must be mapped and writable when this is called. This method will update it as it puts stuff on the kernel stack.

unsafe fn context_switch(
current_kernel_stack: *mut VirtualAddress,
new_kernel_stack: VirtualAddress
)

Do the final part of a context switch: save all the state that needs to be to the current kernel stack, switch to a new kernel stack, and restore all the state from that stack.

unsafe fn drop_into_userspace(kernel_stack_pointer: VirtualAddress) -> !

Do the actual drop into usermode. This assumes that the task's page tables have already been installed, and that an initial frame has been put into the task's kernel stack that this will use to enter userspace.

Loading content...