Module kernel::types

Kernel types.

Structs

ARef

An owned reference to an always-reference-counted object.

Opaque

Stores an opaque value.

ScopeGuard

Runs a cleanup function/closure when dropped.

Enums

Either

A sum type that always holds either a value of type L or R.

Traits

AlwaysRefCounted

Types that are always reference counted.

ForeignOwnable

Used to transfer ownership to and from foreign (non-Rust) languages.