Trait kernel::workqueue::WorkItem

pub trait WorkItem<const ID: u64 = 0> {
    type Pointer: WorkItemPointer<ID>;

    // Required method
    fn run(this: Self::Pointer);
}

Defines the method that should be called when this work item is executed.

This trait is used when the work_struct field is defined using the Work helper.

Required Associated Types

type Pointer: WorkItemPointer<ID>

The pointer type that this struct is wrapped in. This will typically be Arc<Self> or Pin<Box<Self>>.

Required Methods

fn run(this: Self::Pointer)

The method that should be called when this work item is executed.