Module kernel::x86_64

This module defines the kernel entry-point on x86_64.

Re-exports

pub use self::per_cpu::common_per_cpu_data;

pub use self::per_cpu::common_per_cpu_data_mut;

pub use self::task::context_switch;

Modules

acpi_handler

address_space

cpu

interrupts

logger

memory
This module contains the physical memory manager Pebble uses on x86_64.

memory_object

per_cpu

task

Structs

Arch

Statics

ARCH

GDT

Functions

check_support
We rely on certain processor features to be present for simplicity and sanity-retention. This function checks that we support everything we need to, and enable features that need to be.

create_framebuffer

kmain
This is the entry point for the kernel on x86_64. It is called from the UEFI bootloader and initialises the system, then passes control into the common part of the kernel.

load_task