Function kernel::x86_64::task::do_context_switch

unsafe extern "C" fn do_context_switch(
    old_kernel_rsp: *mut VirtualAddress, 
    new_kernel_rsp: VirtualAddress
)

Do the actual context switch: save the context of the old task on its kernel stack, switch to the new task's kernel stack, restore its context and return. The only non-trivial part of this is the returning - for tasks that have run before, we simply work our way back up the kernel callstack and return to userspace from the syscall handler. However, for tasks that have never been run before, the stack frames leading back up to the handler aren't there, and so we manually insert a return to a kernel-space usermode trampoline that enters userspace for the first time in the initial stack frame, which is what the context switch returns to.